% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_hyperbolic_functions.R
\name{acosh_boost}
\alias{acosh_boost}
\title{Inverse Hyperbolic Cosine Function}
\usage{
acosh_boost(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The value of the inverse hyperbolic cosine function acosh at x.
}
\description{
Computes the inverse hyperbolic cosine function acosh(x).
}
