% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airy_functions.R
\name{airy_ai}
\alias{airy_ai}
\title{Airy Function Ai}
\usage{
airy_ai(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The value of the Airy function Ai at x.
}
\description{
Calculates the Airy function Ai(x).
}
