% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airy_functions.R
\name{airy_ai_prime}
\alias{airy_ai_prime}
\title{Airy Function Ai Prime}
\usage{
airy_ai_prime(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The value of the derivative of the Airy function Ai at x.
}
\description{
Calculates the derivative of the Airy function Ai(x).
}
