% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airy_functions.R
\name{airy_ai_zero}
\alias{airy_ai_zero}
\title{Finding Zeros of Airy Function Ai}
\usage{
airy_ai_zero(m)
}
\arguments{
\item{m}{The index of the zero to find (1-based).}
}
\value{
The m-th zero of the Airy function Ai.
}
\description{
Finds the m-th zero of the Airy function Ai.
}
