% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airy_functions.R
\name{airy_bi}
\alias{airy_bi}
\title{Airy Function Bi}
\usage{
airy_bi(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The value of the Airy function Bi at x.
}
\description{
Calculates the Airy function Bi(x).
}
