% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airy_functions.R
\name{airy_bi_prime}
\alias{airy_bi_prime}
\title{Airy Function Bi Prime}
\usage{
airy_bi_prime(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The value of the derivative of the Airy function Bi at x.
}
\description{
Calculates the derivative of the Airy function Bi(x).
}
