% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airy_functions.R
\name{airy_bi_zero}
\alias{airy_bi_zero}
\title{Finding Zeros of Airy Function Bi}
\usage{
airy_bi_zero(m)
}
\arguments{
\item{m}{The index of the zero to find (1-based).}
}
\value{
The m-th zero of the Airy function Bi.
}
\description{
Finds the m-th zero of the Airy function Bi.
}
