% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcsine_distribution.R
\name{arcsine_cdf}
\alias{arcsine_cdf}
\title{Arcsine Distribution Cumulative Distribution Function}
\usage{
arcsine_cdf(x, x_min = 0, x_max = 1)
}
\arguments{
\item{x}{quantile}

\item{x_min}{minimum value of the distribution (default is 0)}

\item{x_max}{maximum value of the distribution (default is 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Arcsine Distribution Cumulative Distribution Function
}
