% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcsine_distribution.R
\name{arcsine_quantile}
\alias{arcsine_quantile}
\title{Arcsine Distribution Quantile Function}
\usage{
arcsine_quantile(p, x_min = 0, x_max = 1)
}
\arguments{
\item{p}{probability}

\item{x_min}{minimum value of the distribution (default is 0)}

\item{x_max}{maximum value of the distribution (default is 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Arcsine Distribution Quantile Function
}
