% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_hyperbolic_functions.R
\name{atanh}
\alias{atanh}
\title{Inverse Hyperbolic Tangent Function}
\usage{
atanh(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The value of the inverse hyperbolic tangent function atanh at x.
}
\description{
Computes the inverse hyperbolic tangent function atanh(x).
}
