% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolators.R
\name{barycentric_rational}
\alias{barycentric_rational}
\title{Barycentric Rational Interpolation}
\usage{
barycentric_rational(x, y, order = 3)
}
\arguments{
\item{x}{Numeric vector of data points (abscissas).}

\item{y}{Numeric vector of data values (ordinates).}

\item{order}{Integer representing the approximation order of the interpolator, defaults to 3.}
}
\value{
An object of class \code{barycentric_rational_interpolator} with methods:
\itemize{
\item \code{interpolate(xi)}: Evaluate the interpolator at point \code{xi}.
\item \code{prime(xi)}: Evaluate the derivative of the interpolator at point \code{xi}.
}
}
\description{
Barycentric rational interpolation is a high-accuracy interpolation method for non-uniformly spaced samples.

\strong{Performance and Accuracy:}

It requires O(N) time for construction and O(N) time for each evaluation.
If the approximation order is d, the error is O(h^(d+1)).

\strong{Caveats:}

This method is robust but can behave unexpectedly if the sample spacing at the endpoints is
much larger than in the center.
}
\examples{
x <- c(0, 1, 2, 3)
y <- c(1, 2, 0, 2)
order <- 3
interpolator <- barycentric_rational(x, y, order)
xi <- 1.5
interpolator$interpolate(xi)
interpolator$prime(xi)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/barycentric.html}{Boost Documentation}
}
