% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_series.R
\name{bernoulli_b2n}
\alias{bernoulli_b2n}
\title{Bernoulli Numbers \eqn{(B_{2n})}}
\usage{
bernoulli_b2n(n)
}
\arguments{
\item{n}{Index of the Bernoulli number (must be a non-negative integer)}
}
\value{
The value of the Bernoulli number \eqn{(B_{2n})}.
}
\description{
Computes the Bernoulli numbers \eqn{(B_{2n})} for even indices.
}
