% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_distribution.R
\name{bernoulli_cdf}
\alias{bernoulli_cdf}
\title{Bernoulli Distribution Cumulative Distribution Function}
\usage{
bernoulli_cdf(x, p)
}
\arguments{
\item{x}{quantile (0 or 1)}

\item{p}{probability of success (0 <= p <= 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Bernoulli Distribution Cumulative Distribution Function
}
