% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_distribution.R
\name{bernoulli_lcdf}
\alias{bernoulli_lcdf}
\title{Bernoulli Distribution Log Cumulative Distribution Function}
\usage{
bernoulli_lcdf(x, p)
}
\arguments{
\item{x}{quantile (0 or 1)}

\item{p}{probability of success (0 <= p <= 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Bernoulli Distribution Log Cumulative Distribution Function
}
