% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_distribution.R
\name{bernoulli_lpdf}
\alias{bernoulli_lpdf}
\title{Bernoulli Distribution Log Probability Density Function}
\usage{
bernoulli_lpdf(x, p)
}
\arguments{
\item{x}{quantile (0 or 1)}

\item{p}{probability of success (0 <= p <= 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Bernoulli Distribution Log Probability Density Function
}
