% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_distribution.R
\name{bernoulli_pdf}
\alias{bernoulli_pdf}
\title{Bernoulli Distribution Probability Density Function}
\usage{
bernoulli_pdf(x, p)
}
\arguments{
\item{x}{quantile (0 or 1)}

\item{p}{probability of success (0 <= p <= 1)}
}
\value{
Probability density function value at \code{x}
}
\description{
Bernoulli Distribution Probability Density Function
}
