% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_distribution.R
\name{bernoulli_quantile}
\alias{bernoulli_quantile}
\title{Bernoulli Distribution Quantile Function}
\usage{
bernoulli_quantile(p, p_success)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{p_success}{probability of success (0 <= p_success <= 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Bernoulli Distribution Quantile Function
}
