% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_functions.R
\name{beta_boost}
\alias{beta_boost}
\title{Euler Beta Function}
\usage{
beta_boost(a, b)
}
\arguments{
\item{a}{First parameter of the beta function}

\item{b}{Second parameter of the beta function}
}
\value{
The value of the Euler beta function B(a, b).
}
\description{
Calculates the Euler beta function B(a, b).
}
