% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distribution.R
\name{beta_cdf}
\alias{beta_cdf}
\title{Beta Distribution Cumulative Distribution Function}
\usage{
beta_cdf(x, alpha, beta)
}
\arguments{
\item{x}{quantile}

\item{alpha}{shape parameter (alpha > 0)}

\item{beta}{shape parameter (beta > 0)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Beta Distribution Cumulative Distribution Function
}
