% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_functions.R
\name{beta_incomplete}
\alias{beta_incomplete}
\title{Full (Non-normalised) Incomplete Beta Function}
\usage{
beta_incomplete(a, b, x)
}
\arguments{
\item{a}{First parameter of the beta function}

\item{b}{Second parameter of the beta function}

\item{x}{Upper limit of integration}
}
\value{
The value of the full incomplete beta function B_x(a, b).
}
\description{
Calculates the full incomplete beta function B_x(a, b).
}
