% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distribution.R
\name{beta_lcdf}
\alias{beta_lcdf}
\title{Beta Distribution Log Cumulative Distribution Function}
\usage{
beta_lcdf(x, alpha, beta)
}
\arguments{
\item{x}{quantile}

\item{alpha}{shape parameter (alpha > 0)}

\item{beta}{shape parameter (beta > 0)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Beta Distribution Log Cumulative Distribution Function
}
