% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distribution.R
\name{beta_lpdf}
\alias{beta_lpdf}
\title{Beta Distribution Log Probability Density Function}
\usage{
beta_lpdf(x, alpha, beta)
}
\arguments{
\item{x}{quantile}

\item{alpha}{shape parameter (alpha > 0)}

\item{beta}{shape parameter (beta > 0)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Beta Distribution Log Probability Density Function
}
