% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distribution.R
\name{beta_pdf}
\alias{beta_pdf}
\title{Beta Distribution Probability Density Function}
\usage{
beta_pdf(x, alpha, beta)
}
\arguments{
\item{x}{quantile}

\item{alpha}{shape parameter (alpha > 0)}

\item{beta}{shape parameter (beta > 0)}
}
\value{
Probability density function value at \code{x}
}
\description{
Beta Distribution Probability Density Function
}
