% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_distribution.R
\name{beta_quantile}
\alias{beta_quantile}
\title{Beta Distribution Quantile Function}
\usage{
beta_quantile(p, alpha, beta)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{alpha}{shape parameter (alpha > 0)}

\item{beta}{shape parameter (beta > 0)}
}
\value{
Quantile function value at \code{p}
}
\description{
Beta Distribution Quantile Function
}
