% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_functions.R
\name{betac}
\alias{betac}
\title{Full (Non-normalised) Complement of the Incomplete Beta Function}
\usage{
betac(a, b, x)
}
\arguments{
\item{a}{First parameter of the beta function}

\item{b}{Second parameter of the beta function}

\item{x}{Upper limit of integration}
}
\value{
The value of the full complement of the incomplete beta function 1 - B_x(a, b).
}
\description{
Calculates the full complement of the incomplete beta function 1 - B_x(a, b).
}
