% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolators.R
\name{bezier_polynomial}
\alias{bezier_polynomial}
\title{Bezier Polynomial Interpolator}
\usage{
bezier_polynomial(control_points)
}
\arguments{
\item{control_points}{List of control points, where each element is a numeric vector of length 3.}
}
\value{
An object of class \code{bezier_polynomial} with methods:
\itemize{
\item \code{interpolate(xi)}: Evaluate the interpolator at point \code{xi}.
\item \code{prime(xi)}: Evaluate the derivative of the interpolator at point \code{xi}.
\item \code{edit_control_point(new_control_point, index)}: Insert a new control point at the specified index.
}
}
\description{
Bezier polynomials are smooth curves which approximate a set of control points.
They are commonly used in computer-aided geometric design.

\strong{Properties:}

The curve is approximating, meaning it does not necessarily pass through the control points.
Passing n control points creates a polynomial of degree n-1.
Evaluation is O(N^2) via de Casteljau's algorithm.
}
\examples{
control_points <- list(c(0, 0, 0), c(1, 2, 0), c(2, 0, 0), c(3, 3, 0))
interpolator <- bezier_polynomial(control_points)
xi <- 1.5
interpolator$interpolate(xi)
interpolator$prime(xi)
new_control_point <- c(1.5, 1, 0)
interpolator$edit_control_point(new_control_point, 2)

}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/bezier_polynomial.html}{Boost Documentation}
}
