% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_distribution.R
\name{binomial_cdf}
\alias{binomial_cdf}
\title{Binomial Cumulative Distribution Function}
\usage{
binomial_cdf(k, n, prob)
}
\arguments{
\item{k}{number of successes (0 <= k <= n)}

\item{n}{number of trials (n >= 0)}

\item{prob}{probability of success on each trial (0 <= prob <= 1)}
}
\value{
Cumulative distribution function value at \code{k}
}
\description{
Binomial Cumulative Distribution Function
}
\examples{
binomial_cdf(3, 10, 0.5)
}
