% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorials_and_binomial_coefficients.R
\name{binomial_coefficient}
\alias{binomial_coefficient}
\title{Binomial Coefficient}
\usage{
binomial_coefficient(n, k)
}
\arguments{
\item{n}{Total number of elements}

\item{k}{Number of elements to choose}
}
\value{
The binomial coefficient C(n, k).
}
\description{
Computes the binomial coefficient "n choose k", which is the number of ways to choose k elements from a set of n elements.
}
