% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_distribution.R
\name{binomial_lcdf}
\alias{binomial_lcdf}
\title{Binomial Log Cumulative Distribution Function}
\usage{
binomial_lcdf(k, n, prob)
}
\arguments{
\item{k}{number of successes (0 <= k <= n)}

\item{n}{number of trials (n >= 0)}

\item{prob}{probability of success on each trial (0 <= prob <= 1)}
}
\value{
Log cumulative distribution function value at \code{k}
}
\description{
Binomial Log Cumulative Distribution Function
}
\examples{
binomial_lcdf(3, 10, 0.5)
}
