% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_distribution.R
\name{binomial_pdf}
\alias{binomial_pdf}
\title{Binomial Probability Density Function}
\usage{
binomial_pdf(k, n, prob)
}
\arguments{
\item{k}{number of successes (0 <= k <= n)}

\item{n}{number of trials (n >= 0)}

\item{prob}{probability of success on each trial (0 <= prob <= 1)}
}
\value{
Probability density function value at \code{k}
}
\description{
Binomial Probability Density Function
}
\examples{
binomial_pdf(3, 10, 0.5)
}
