% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_distribution.R
\name{binomial_quantile}
\alias{binomial_quantile}
\title{Binomial Quantile Function}
\usage{
binomial_quantile(p, n, prob)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{n}{number of trials (n >= 0)}

\item{prob}{probability of success on each trial (0 <= prob <= 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Binomial Quantile Function
}
\examples{
binomial_quantile(0.5, 10, 0.5)
}
