% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_hyperbolic_functions.R
\name{boost_acosh}
\alias{boost_acosh}
\title{Inverse Hyperbolic Cosine Function}
\usage{
boost_acosh(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The value of the inverse hyperbolic cosine function acosh at x.
}
\description{
Computes the inverse hyperbolic cosine function acosh(x).
}
