% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_hyperbolic_functions.R
\name{boost_asinh}
\alias{boost_asinh}
\title{Inverse Hyperbolic Sine Function}
\usage{
boost_asinh(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The value of the inverse hyperbolic sine function asinh at x.
}
\description{
Computes the inverse hyperbolic sine function asinh(x).
}
