% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_functions.R
\name{boost_beta}
\alias{boost_beta}
\title{Euler Beta Function}
\usage{
boost_beta(a, b)
}
\arguments{
\item{a}{First parameter of the beta function}

\item{b}{Second parameter of the beta function}
}
\value{
The value of the Euler beta function B(a, b).
}
\description{
Calculates the Euler beta function B(a, b).
}
