% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{boost_expm1}
\alias{boost_expm1}
\title{Exponential of x and substract 1}
\usage{
boost_expm1(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The exponential of x minus 1.
}
\description{
Computes the exponential of x and subtracts 1.
}
