% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{boost_lgamma}
\alias{boost_lgamma}
\title{Logarithm of the Gamma Function}
\usage{
boost_lgamma(z)
}
\arguments{
\item{z}{Input numeric value}
}
\value{
The logarithm of the gamma function at z.
}
\description{
Computes the logarithm of the gamma function for a given input.
}
