% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{boost_log1p}
\alias{boost_log1p}
\title{Logarithm of 1 plus x}
\usage{
boost_log1p(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The natural logarithm of 1 plus x.
}
\description{
Computes the natural logarithm of 1 plus x.
}
