% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostmath-package.R
\docType{package}
\name{boostmath-package}
\alias{boostmath}
\alias{boostmath-package}
\title{boostmath: Exposing Boost's Math Functions for R}
\description{
'R' bindings for the various functions and statistical distributions provided by the Boost Math library.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrjohns/boostmath}
  \item Report bugs at \url{https://github.com/andrjohns/boostmath/issues}
}

}
\author{
\strong{Maintainer}: Andrew R. Johnson \email{andrew.johnson@arjohnsonau.com} (\href{https://orcid.org/0000-0001-7000-8065}{ORCID})

}
\keyword{internal}
