% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootfinding_and_minimisation.R
\name{bracket_and_solve_root}
\alias{bracket_and_solve_root}
\title{Root-Finding with Bracket and Solve Method}
\usage{
bracket_and_solve_root(
  f,
  guess,
  factor,
  rising,
  digits = .Machine$double.digits,
  max_iter = .Machine$integer.max
)
}
\arguments{
\item{f}{A function to find the root of. It should take and return a single numeric value.}

\item{guess}{A numeric value that is a guess for the root.}

\item{factor}{Size of steps to take when searching}

\item{rising}{If TRUE, the function is assumed to be rising, otherwise it is assumed to be falling.}

\item{digits}{The number of significant digits to which the root should be found. Defaults to double precision.}

\item{max_iter}{The maximum number of iterations to perform. Defaults to the maximum integer value.}
}
\value{
A three-element list containing:
\itemize{
\item \code{lower}: The lower bound of the interval containing the root.
\item \code{upper}: The upper bound of the interval containing the root.
\item \code{iterations}: The number of iterations performed.
}
}
\description{
Root-Finding with Bracket and Solve Method
}
\examples{
f <- function(x) x^2 - 2
bracket_and_solve_root(f, guess = 1, factor = 0.1, rising = TRUE)
}
