% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootfinding_and_minimisation.R
\name{brent_find_minima}
\alias{brent_find_minima}
\title{Finding Minima using Brent's Method}
\usage{
brent_find_minima(
  f,
  a,
  b,
  digits = .Machine$double.digits,
  max_iter = .Machine$integer.max
)
}
\arguments{
\item{f}{A function to find the minimum of. It should take and return a single numeric value.}

\item{a}{The lower bound of the interval to search for the minimum.}

\item{b}{The upper bound of the interval to search for the minimum.}

\item{digits}{The number of significant digits to which the minimum should be found. Defaults to double precision.}

\item{max_iter}{The maximum number of iterations to perform. Defaults to the maximum integer value.}
}
\value{
A three-element list containing:
\itemize{
\item \code{minimum}: The value which minimises the function.
\item \code{value}: The value of the function at the minimum.
\item \code{iterations}: The number of iterations performed.
}
}
\description{
Finding Minima using Brent's Method
}
\examples{
f <- function(x) (x - 2)^2 + 1
brent_find_minima(f, a = 0, b = 4)
}
