% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolators.R
\name{cardinal_cubic_hermite}
\alias{cardinal_cubic_hermite}
\title{Cardinal Cubic Hermite Interpolator}
\usage{
cardinal_cubic_hermite(y, dydx, x0, dx)
}
\arguments{
\item{y}{Numeric vector of ordinates (y-coordinates).}

\item{dydx}{Numeric vector of derivatives (slopes) at each point.}

\item{x0}{Numeric value of the first abscissa (x-coordinate).}

\item{dx}{Numeric value of the spacing between abscissas.}
}
\value{
An object of class \code{cardinal_cubic_hermite} with methods:
\itemize{
\item \code{interpolate(xi)}: Evaluate the interpolator at point \code{xi}.
\item \code{prime(xi)}: Evaluate the derivative of the interpolator at point \code{xi}.
\item \code{domain()}: Get the domain of the interpolator.
}
}
\description{
The cardinal cubic Hermite interpolator is similar to the cubic Hermite interpolator but optimized for equispaced data.

\strong{Performance:}

This allows for constant-time evaluation.
}
\examples{
y <- c(0, 1, 0)
dydx <- c(1, 0, -1)
interpolator <- cardinal_cubic_hermite(y, dydx, 0, 1)
xi <- 0.5
interpolator$interpolate(xi)
interpolator$prime(xi)
interpolator$domain()
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/cubic_hermite.html}{Boost Documentation}
}
