% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_distribution.R
\name{cauchy_cdf}
\alias{cauchy_cdf}
\title{Cauchy Cumulative Distribution Function}
\usage{
cauchy_cdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Cauchy Cumulative Distribution Function
}
\examples{
cauchy_cdf(0)
}
