% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_distribution.R
\name{cauchy_lcdf}
\alias{cauchy_lcdf}
\title{Cauchy Log Cumulative Distribution Function}
\usage{
cauchy_lcdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Cauchy Log Cumulative Distribution Function
}
\examples{
cauchy_lcdf(0)
}
