% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_distribution.R
\name{cauchy_lpdf}
\alias{cauchy_lpdf}
\title{Cauchy Log Probability Density Function}
\usage{
cauchy_lpdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Cauchy Log Probability Density Function
}
\examples{
cauchy_lpdf(0)
}
