% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cauchy_distribution.R
\name{cauchy_pdf}
\alias{cauchy_pdf}
\title{Cauchy Probability Density Function}
\usage{
cauchy_pdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Probability density function value at \code{x}
}
\description{
Cauchy Probability Density Function
}
\examples{
cauchy_pdf(0)
}
