% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{chebyshev_clenshaw_recurrence}
\alias{chebyshev_clenshaw_recurrence}
\title{Next Chebyshev Polynomial using Clenshaw's Recurrence}
\usage{
chebyshev_clenshaw_recurrence(c, x)
}
\arguments{
\item{c}{Coefficients of the Chebyshev polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the next Chebyshev polynomial \eqn{(T_{n+1}(x))}.
}
\description{
Computes the next Chebyshev polynomial of the first kind \eqn{(T_{n+1}(x))} using Clenshaw's recurrence relation.
}
