% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{chebyshev_clenshaw_recurrence_ab}
\alias{chebyshev_clenshaw_recurrence_ab}
\title{Chebyshev Polynomial of the First Kind using Clenshaw's Recurrence with Interval \link{a, b}}
\usage{
chebyshev_clenshaw_recurrence_ab(c, a, b, x)
}
\arguments{
\item{c}{Coefficients of the Chebyshev polynomial}

\item{a}{Lower bound of the interval}

\item{b}{Upper bound of the interval}

\item{x}{Argument of the polynomial}
}
\value{
The value of the Chebyshev polynomial \(T_n(x)\) evaluated at \(x\) in the interval \(\link{a, b}\).
}
\description{
Computes the Chebyshev polynomial of the first kind \(T_n(x)\) using Clenshaw's recurrence relation with a specified interval \(\link{a, b}\).
}
