% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{chebyshev_next}
\alias{chebyshev_next}
\title{Next Chebyshev Polynomial of the First Kind}
\usage{
chebyshev_next(x, Tn, Tn_1)
}
\arguments{
\item{x}{Argument of the polynomial}

\item{Tn}{Value of the Chebyshev polynomial \eqn{\(T_n(x)\)}}

\item{Tn_1}{Value of the Chebyshev polynomial \eqn{\(T_{n-1}(x)\)}}
}
\value{
The value of the next Chebyshev polynomial \eqn{\(T_{n+1}(x)\)}.
}
\description{
Computes the next Chebyshev polynomial of the first kind \eqn{\(T_{n+1}(x)\)} using the previous two polynomials.
}
