% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{chebyshev_t_prime}
\alias{chebyshev_t_prime}
\title{Chebyshev Polynomial of the First Kind (Derivative)}
\usage{
chebyshev_t_prime(n, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the derivative of the Chebyshev polynomial \eqn{\(T_n'(x)\)}.
}
\description{
Computes the derivative of the Chebyshev polynomial of the first kind \eqn{\(T_n'(x)\)}.
}
