% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{chebyshev_u}
\alias{chebyshev_u}
\title{Chebyshev Polynomial of the Second Kind}
\usage{
chebyshev_u(n, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the Chebyshev polynomial \eqn{\(U_n(x)\)}.
}
\description{
Computes the Chebyshev polynomial of the second kind \eqn{\(U_n(x)\)}.
}
