% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_distribution.R
\name{chi_squared_lcdf}
\alias{chi_squared_lcdf}
\title{Chi-Squared Log Cumulative Distribution Function}
\usage{
chi_squared_lcdf(x, df)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (df > 0)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Chi-Squared Log Cumulative Distribution Function
}
\examples{
chi_squared_lcdf(2, 3)
}
