% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_distribution.R
\name{chi_squared_lpdf}
\alias{chi_squared_lpdf}
\title{Chi-Squared Log Probability Density Function}
\usage{
chi_squared_lpdf(x, df)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (df > 0)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Chi-Squared Log Probability Density Function
}
\examples{
chi_squared_lpdf(2, 3)
}
