% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_distribution.R
\name{chi_squared_pdf}
\alias{chi_squared_pdf}
\title{Chi-Squared Probability Density Function}
\usage{
chi_squared_pdf(x, df)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (df > 0)}
}
\value{
Probability density function value at \code{x}
}
\description{
Chi-Squared Probability Density Function
}
\examples{
chi_squared_pdf(2, 3)
}
