% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_squared_distribution.R
\name{chi_squared_quantile}
\alias{chi_squared_quantile}
\title{Chi-Squared Quantile Function}
\usage{
chi_squared_quantile(p, df)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{df}{degrees of freedom (df > 0)}
}
\value{
Quantile function value at \code{p}
}
\description{
Chi-Squared Quantile Function
}
\examples{
chi_squared_quantile(0.5, 3)
}
