% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature_and_differentiation.R
\name{complex_step_derivative}
\alias{complex_step_derivative}
\title{Complex Step Numerical Differentiation}
\usage{
complex_step_derivative(f, x)
}
\arguments{
\item{f}{A function to differentiate. It should accept a single complex value and
return a single complex value.}

\item{x}{The point at which to evaluate the derivative, as a numeric (not complex) value.}
}
\value{
The approximate value of the derivative at the point \code{x}.
}
\description{
This function computes the numerical derivative of a function using the complex step method.
}
\examples{
complex_step_derivative(exp, 1.7)
}
