% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\title{Boost Math Constants}
\usage{
constants()
}
\value{
A list of constants (see documentation below for full list).
}
\description{
Provides access to mathematical constants used in the Boost Math library.
}
\examples{
constants()
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/constants.html}{Boost Documentation} for more details on the constants.
}
