% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{cos_pi}
\alias{cos_pi}
\title{Cosine of Pi}
\usage{
cos_pi(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The cosine of x multiplied by pi.
}
\description{
Computes the cosine of a value multiplied by pi.
}
