% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootfinding_and_minimisation.R
\name{cubic_root_condition_number}
\alias{cubic_root_condition_number}
\title{Finding Condition Number of a Cubic Root}
\usage{
cubic_root_condition_number(a, b, c, d, root)
}
\arguments{
\item{a}{Coefficient of the cubic term.}

\item{b}{Coefficient of the quadratic term.}

\item{c}{Coefficient of the linear term.}

\item{d}{Constant term.}

\item{root}{The root to evaluate the condition number at.}
}
\value{
The condition number of the cubic equation at the specified root.
}
\description{
Finding Condition Number of a Cubic Root
}
\examples{
cubic_root_condition_number(1, -6, 11, -6, 1)
}
