% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootfinding_and_minimisation.R
\name{cubic_root_residual}
\alias{cubic_root_residual}
\title{Finding Residual of a Cubic Root}
\usage{
cubic_root_residual(a, b, c, d, root)
}
\arguments{
\item{a}{Coefficient of the cubic term.}

\item{b}{Coefficient of the quadratic term.}

\item{c}{Coefficient of the linear term.}

\item{d}{Constant term.}

\item{root}{The root to evaluate the residual at.}
}
\value{
The residual of the cubic equation at the specified root.
}
\description{
Finding Residual of a Cubic Root
}
\examples{
cubic_root_residual(1, -6, 11, -6, 1)
}
