% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootfinding_and_minimisation.R
\name{cubic_roots}
\alias{cubic_roots}
\title{Finding Cubic Roots}
\usage{
cubic_roots(a, b, c, d)
}
\arguments{
\item{a}{Coefficient of the cubic term.}

\item{b}{Coefficient of the quadratic term.}

\item{c}{Coefficient of the linear term.}

\item{d}{Constant term.}
}
\value{
A numeric vector containing the three roots of the cubic equation.
}
\description{
Finding Cubic Roots
}
\examples{
cubic_roots(1, -6, 11, -6)
}
