% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_functions.R
\name{cyl_bessel_j}
\alias{cyl_bessel_j}
\title{Bessel Function of the First Kind}
\usage{
cyl_bessel_j(v, x)
}
\arguments{
\item{v}{Order of the Bessel function}

\item{x}{Argument of the Bessel function}
}
\value{
The value of the Bessel function of the first kind J_v(x).
}
\description{
Calculates the Bessel function of the first kind J_v(x).
}
