% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_functions.R
\name{cyl_bessel_j_prime}
\alias{cyl_bessel_j_prime}
\title{Derivative of the Bessel Function of the First Kind}
\usage{
cyl_bessel_j_prime(v, x)
}
\arguments{
\item{v}{Order of the Bessel function}

\item{x}{Argument of the Bessel function}
}
\value{
The value of the derivative of the Bessel function of the first kind J_v(x).
}
\description{
Calculates the derivative of the Bessel function of the first kind J_v(x).
}
