% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_functions.R
\name{cyl_bessel_j_zero}
\alias{cyl_bessel_j_zero}
\title{Finding Zeros of Bessel Function of the First Kind}
\usage{
cyl_bessel_j_zero(v, m)
}
\arguments{
\item{v}{Order of the Bessel function}

\item{m}{The index of the zero to find (1-based).}
}
\value{
The m-th zero of the Bessel function of the first kind J_v(x).
}
\description{
Finds the m-th zero of the Bessel function of the first kind J_v(x).
}
