% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_functions.R
\name{cyl_bessel_j_zero_multiple}
\alias{cyl_bessel_j_zero_multiple}
\title{Finding Multiple Zeros of Bessel Function of the First Kind}
\usage{
cyl_bessel_j_zero_multiple(v, start_index, number_of_zeros)
}
\arguments{
\item{v}{Order of the Bessel function}

\item{start_index}{The starting index for the zeros (1-based).}

\item{number_of_zeros}{The number of zeros to find.}
}
\value{
A numeric vector containing the zeros of the Bessel function of the
first kind J_v(x).
}
\description{
Finds multiple zeros of the Bessel function of the first kind J_v(x) starting
from a given index.
}
