% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_functions.R
\name{cyl_bessel_k_prime}
\alias{cyl_bessel_k_prime}
\title{Derivative of the Modified Bessel Function of the Second Kind}
\usage{
cyl_bessel_k_prime(v, x)
}
\arguments{
\item{v}{Order of the modified Bessel function}

\item{x}{Argument of the modified Bessel function}
}
\value{
The value of the derivative of the modified Bessel function of the second kind K_v(x).
}
\description{
Calculates the derivative of the modified Bessel function of the second kind K_v(x).
}
