% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_functions.R
\name{cyl_neumann}
\alias{cyl_neumann}
\title{Bessel Function of the Second Kind}
\usage{
cyl_neumann(v, x)
}
\arguments{
\item{v}{Order of the Bessel function}

\item{x}{Argument of the Bessel function}
}
\value{
The value of the Bessel function of the second kind Y_v(x).
}
\description{
Calculates the Bessel function of the second kind Y_v(x).
}
