% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_functions.R
\name{cyl_neumann_zero}
\alias{cyl_neumann_zero}
\title{Finding Zeros of Bessel Function of the Second Kind}
\usage{
cyl_neumann_zero(v, m)
}
\arguments{
\item{v}{Order of the Bessel function}

\item{m}{The index of the zero to find (1-based).}
}
\value{
The m-th zero of the Bessel function of the second kind Y_v
}
\description{
Finds the m-th zero of the Bessel function of the second kind Y_v(x).
}
