% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{digamma_boost}
\alias{digamma_boost}
\title{Digamma Function}
\usage{
digamma_boost(z)
}
\arguments{
\item{z}{Input numeric value}
}
\value{
The value of the digamma function at z.
}
\description{
Computes the digamma function (the logarithmic derivative of the gamma function) for a given input.
}
