% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorials_and_binomial_coefficients.R
\name{double_factorial}
\alias{double_factorial}
\title{Double Factorial}
\usage{
double_factorial(i)
}
\arguments{
\item{i}{Non-negative integer input}
}
\value{
The double factorial of the input integer.
}
\description{
Computes the double factorial of a non-negative integer.
}
