% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_integrals.R
\name{ellint_1}
\alias{ellint_1}
\title{Complete Elliptic Integral of the First Kind}
\usage{
ellint_1(k)
}
\arguments{
\item{k}{Elliptic modulus}
}
\value{
The value of the complete elliptic integral of the first kind.
}
\description{
Computes the complete elliptic integral of the first kind.
}
