% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_integrals.R
\name{ellint_1_incomplete}
\alias{ellint_1_incomplete}
\title{Incomplete Elliptic Integral of the First Kind - Legendre's Form}
\usage{
ellint_1_incomplete(k, phi)
}
\arguments{
\item{k}{Elliptic modulus}

\item{phi}{Amplitude}
}
\value{
The value of the incomplete elliptic integral of the first kind.
}
\description{
Computes the incomplete elliptic integral of the first kind in Legendre's form.
}
