% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_integrals.R
\name{ellint_2}
\alias{ellint_2}
\title{Complete Elliptic Integral of the Second Kind}
\usage{
ellint_2(k)
}
\arguments{
\item{k}{Elliptic modulus}
}
\value{
The value of the complete elliptic integral of the second kind.
}
\description{
Computes the complete elliptic integral of the second kind.
}
