% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_integrals.R
\name{ellint_2_incomplete}
\alias{ellint_2_incomplete}
\title{Incomplete Elliptic Integral of the Second Kind - Legendre's Form}
\usage{
ellint_2_incomplete(k, phi)
}
\arguments{
\item{k}{Elliptic modulus}

\item{phi}{Amplitude}
}
\value{
The value of the incomplete elliptic integral of the second kind.
}
\description{
Computes the incomplete elliptic integral of the second kind in Legendre's form.
}
