% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_integrals.R
\name{ellint_3}
\alias{ellint_3}
\title{Complete Elliptic Integral of the Third Kind}
\usage{
ellint_3(k, n)
}
\arguments{
\item{k}{Elliptic modulus}

\item{n}{Characteristic}
}
\value{
The value of the complete elliptic integral of the third kind.
}
\description{
Computes the complete elliptic integral of the third kind.
}
