% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_integrals.R
\name{ellint_3_incomplete}
\alias{ellint_3_incomplete}
\title{Incomplete Elliptic Integral of the Third Kind - Legendre's Form}
\usage{
ellint_3_incomplete(k, n, phi)
}
\arguments{
\item{k}{Elliptic modulus}

\item{n}{Characteristic}

\item{phi}{Amplitude}
}
\value{
The value of the incomplete elliptic integral of the third kind.
}
\description{
Computes the incomplete elliptic integral of the third kind in Legendre's form.
}
