% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_integrals.R
\name{ellint_d}
\alias{ellint_d}
\title{Complete Elliptic Integral D}
\usage{
ellint_d(k)
}
\arguments{
\item{k}{Elliptic modulus}
}
\value{
The value of the complete elliptic integral D.
}
\description{
Computes the complete elliptic integral D.
}
