% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_integrals.R
\name{ellint_d_incomplete}
\alias{ellint_d_incomplete}
\title{Incomplete Elliptic Integral D}
\usage{
ellint_d_incomplete(k, phi)
}
\arguments{
\item{k}{Elliptic modulus}

\item{phi}{Amplitude}
}
\value{
The value of the incomplete elliptic integral D.
}
\description{
Computes the incomplete elliptic integral D in Legendre's form.
}
