% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_integrals.R
\name{ellint_rc}
\alias{ellint_rc}
\title{Carlson's Elliptic Integral Rc}
\usage{
ellint_rc(x, y)
}
\arguments{
\item{x}{First parameter of the integral}

\item{y}{Second parameter of the integral}

\item{z}{Third parameter of the integral}
}
\value{
The value of Carlson's elliptic integral Rc.
}
\description{
Computes Carlson's elliptic integral Rc.
}
