% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_integrals.R
\name{ellint_rg}
\alias{ellint_rg}
\title{Carlson's Elliptic Integral Rg}
\usage{
ellint_rg(x, y, z)
}
\arguments{
\item{x}{First parameter of the integral}

\item{y}{Second parameter of the integral}

\item{z}{Third parameter of the integral}
}
\value{
The value of Carlson's elliptic integral Rg.
}
\description{
Computes Carlson's elliptic integral Rg.
}
