% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_integrals.R
\name{ellint_rj}
\alias{ellint_rj}
\title{Carlson's Elliptic Integral Rj}
\usage{
ellint_rj(x, y, z, p)
}
\arguments{
\item{x}{First parameter of the integral}

\item{y}{Second parameter of the integral}

\item{z}{Third parameter of the integral}

\item{p}{Fourth parameter of the integral}
}
\value{
The value of Carlson's elliptic integral Rj.
}
\description{
Computes Carlson's elliptic integral Rj.
}
