% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{erf}
\alias{erf}
\title{Error Function erf}
\usage{
erf(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The value of the error function erf at x.
}
\description{
Computes the error function erf(x).
}
