% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{erf_inv}
\alias{erf_inv}
\title{Inverse of the Error Function erf}
\usage{
erf_inv(p)
}
\arguments{
\item{p}{Probability value}
}
\value{
The value of x such that erf(x) = p.
}
\description{
Computes the inverse of the error function erf
}
