% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{erfc}
\alias{erfc}
\title{Complementary Error Function erfc}
\usage{
erfc(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The value of the complementary error function erfc at x.
}
\description{
Computes the complementary error function erfc(x).
}
