% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{erfc_inv}
\alias{erfc_inv}
\title{Inverse of the Complementary Error Function erfc}
\usage{
erfc_inv(p)
}
\arguments{
\item{p}{Probability value}
}
\value{
The value of x such that erfc(x) = p.
}
\description{
Computes the inverse of the complementary error function erfc
}
