% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{error_functions}
\alias{error_functions}
\alias{erf}
\alias{erfc}
\alias{erf_inv}
\alias{erfc_inv}
\title{Error Functions and Inverses}
\usage{
erf(x)

erfc(x)

erf_inv(p)

erfc_inv(p)
}
\arguments{
\item{x}{Input numeric value}

\item{p}{Probability value (0 <= p <= 1)}
}
\value{
A single numeric value with the computed error function, complementary error function, or their inverses.
}
\description{
Functions to compute the error function, complementary error function, and their inverses.

\strong{Error Function erf and complement erfc}
\itemize{
\item \code{erf(x)}: Error function \eqn{erf(x)}:
}

\deqn{erf(x) = \frac{2}{\sqrt{\pi}}\int_{0}^xe^{-t^2}dt}
\itemize{
\item \code{erfc(x)}: Error function complement \eqn{1 - erf(x)}
}

\strong{Error Function Inverses}
\itemize{
\item \code{erf_inv(p)}: Inverse of the error function \eqn{p = erf(x)}:
\item \code{erfx_inv(p)}: Inverse of the error function complement \eqn{p = erfc(x)}:
}
}
\examples{
# Error function
erf(0.5)
# Complementary error function
erfc(0.5)
# Inverse error function
erf_inv(0.5)
# Inverse complementary error function
erfc_inv(0.5)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/sf_erf.html}{Boost Documentation} for more details
}
