% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature_and_differentiation.R
\name{exp_sinh}
\alias{exp_sinh}
\title{Exp-Sinh Quadrature for Numerical Integration}
\usage{
exp_sinh(f, a, b, tol = .Machine$double.eps, max_refinements = 9)
}
\arguments{
\item{f}{A function to integrate. It should accept a single numeric value and
return a single numeric value.}

\item{a}{The lower limit of integration.}

\item{b}{The upper limit of integration.}

\item{tol}{The tolerance for the approximation. Default is \code{.Machine$double.eps}.}

\item{max_refinements}{The maximum number of refinements to apply. Default is 9.}
}
\value{
The approximate value of the integral.
}
\description{
This function performs integration using an exp-sinh quadrature routine.
}
