% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_integrals.R
\name{expint_ei}
\alias{expint_ei}
\title{Exponential Integral Ei}
\usage{
expint_ei(z)
}
\arguments{
\item{z}{Argument of the integral}
}
\value{
The value of the exponential integral Ei.
}
\description{
Computes the exponential integral Ei.
}
