% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_integrals.R
\name{expint_en}
\alias{expint_en}
\title{Exponential Integral En}
\usage{
expint_en(n, z)
}
\arguments{
\item{n}{Order of the integral}

\item{z}{Argument of the integral}
}
\value{
The value of the exponential integral En.
}
\description{
Computes the exponential integral En.
}
