% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{expm1_boost}
\alias{expm1_boost}
\title{Exponential of x and substract 1}
\usage{
expm1_boost(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The exponential of x minus 1.
}
\description{
Computes the exponential of x and subtracts 1.
}
