% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_distribution.R
\name{exponential_cdf}
\alias{exponential_cdf}
\title{Exponential Cumulative Distribution Function}
\usage{
exponential_cdf(x, lambda)
}
\arguments{
\item{x}{quantile}

\item{lambda}{rate parameter (lambda > 0)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Exponential Cumulative Distribution Function
}
\examples{
exponential_cdf(1, 2)
}
