% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_distribution.R
\name{exponential_lcdf}
\alias{exponential_lcdf}
\title{Exponential Log Cumulative Distribution Function}
\usage{
exponential_lcdf(x, lambda)
}
\arguments{
\item{x}{quantile}

\item{lambda}{rate parameter (lambda > 0)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Exponential Log Cumulative Distribution Function
}
\examples{
exponential_lcdf(1, 2)
}
