% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_distribution.R
\name{exponential_lpdf}
\alias{exponential_lpdf}
\title{Exponential Log Probability Density Function}
\usage{
exponential_lpdf(x, lambda)
}
\arguments{
\item{x}{quantile}

\item{lambda}{rate parameter (lambda > 0)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Exponential Log Probability Density Function
}
\examples{
exponential_lpdf(1, 2)
}
