% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_distribution.R
\name{exponential_pdf}
\alias{exponential_pdf}
\title{Exponential Probability Density Function}
\usage{
exponential_pdf(x, lambda)
}
\arguments{
\item{x}{quantile}

\item{lambda}{rate parameter (lambda > 0)}
}
\value{
Probability density function value at \code{x}
}
\description{
Exponential Probability Density Function
}
\examples{
exponential_pdf(1, 2)
}
