% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_distribution.R
\name{exponential_quantile}
\alias{exponential_quantile}
\title{Exponential Quantile Function}
\usage{
exponential_quantile(p, lambda)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{lambda}{rate parameter (lambda > 0)}
}
\value{
Quantile function value at \code{p}
}
\description{
Exponential Quantile Function
}
\examples{
exponential_quantile(0.5, 2)
}
