% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_value_distribution.R
\name{extreme_value_lcdf}
\alias{extreme_value_lcdf}
\title{Extreme Value Log Cumulative Distribution Function}
\usage{
extreme_value_lcdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Extreme Value Log Cumulative Distribution Function
}
\examples{
extreme_value_lcdf(0)
}
