% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_value_distribution.R
\name{extreme_value_lpdf}
\alias{extreme_value_lpdf}
\title{Extreme Value Log Probability Density Function}
\usage{
extreme_value_lpdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Extreme Value Log Probability Density Function
}
\examples{
extreme_value_lpdf(0)
}
