% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extreme_value_distribution.R
\name{extreme_value_quantile}
\alias{extreme_value_quantile}
\title{Extreme Value Quantile Function}
\usage{
extreme_value_quantile(p, location = 0, scale = 1)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Extreme Value Quantile Function
}
\examples{
extreme_value_quantile(0.5)
}
