% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorials_and_binomial_coefficients.R
\name{factorial}
\alias{factorial}
\title{Factorial}
\usage{
factorial(i)
}
\arguments{
\item{i}{Non-negative integer input}
}
\value{
The factorial of the input integer.
}
\description{
Computes the factorial of a non-negative integer.
}
