% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorials_and_binomial_coefficients.R
\name{falling_factorial}
\alias{falling_factorial}
\title{Falling Factorial}
\usage{
falling_factorial(x, i)
}
\arguments{
\item{x}{Base value}

\item{i}{Exponent (non-negative integer)}
}
\value{
The value of the falling factorial.
}
\description{
Computes the falling factorial of a number.
}
