% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filters}
\alias{filters}
\alias{daubechies_scaling_filter}
\alias{daubechies_wavelet_filter}
\title{Filters}
\usage{
daubechies_scaling_filter(order)

daubechies_wavelet_filter(order)
}
\arguments{
\item{order}{An integer specifying the number of vanishing moments (1 to 19).}
}
\value{
A numeric vector of size 2*order containing the filter coefficients.
}
\description{
Functions to compute Daubechies scaling and wavelet filter coefficients.

The returned coefficients correspond to the compactly supported Daubechies
wavelets indexed by the number of vanishing moments $p$, returning $2p$ taps.

\strong{Conventions:}
Boost indexes filters by vanishing moments (as in PyWavelets and Mathematica),
normalizes coefficients to unit \eqn{\ell_2} norm, and uses the convolutional ordering
shown in Daubechies (1988). Other libraries may index by number of taps, use a
\eqn{\sqrt{2}} scaling, or reverse coefficient order.
}
\examples{
# Daubechies Scaling Filter of order 4
daubechies_scaling_filter(4)
# Daubechies Wavelet Filter of order 4
daubechies_wavelet_filter(4)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/daubechies_filters.html}{Boost Documentation}
for more details on the mathematical background.
}
