% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature_and_differentiation.R
\name{finite_difference_derivative}
\alias{finite_difference_derivative}
\title{Finite Difference Numerical Differentiation}
\usage{
finite_difference_derivative(f, x, order = 1)
}
\arguments{
\item{f}{A function to differentiate. It should accept a single numeric value and}

\item{x}{The point at which to evaluate the derivative.}

\item{order}{The order of accuracy of the derivative to compute. Default is 1.}
}
\value{
The approximate value of the derivative at the point \code{x}.
}
\description{
This function computes the numerical derivative of a function using finite differences.
}
