% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_f_distribution.R
\name{fisher_f_pdf}
\alias{fisher_f_pdf}
\title{Fisher F Distribution Probability Density Function}
\usage{
fisher_f_pdf(x, df1, df2)
}
\arguments{
\item{x}{quantile}

\item{df1}{degrees of freedom for the numerator (df1 > 0)}

\item{df2}{degrees of freedom for the denominator (df2 > 0)}
}
\value{
Probability density function value at \code{x}
}
\description{
Fisher F Distribution Probability Density Function
}
\examples{
fisher_f_pdf(1, 5, 2)
}
