% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_f_distribution.R
\name{fisher_f_quantile}
\alias{fisher_f_quantile}
\title{Fisher F Distribution Quantile Function}
\usage{
fisher_f_quantile(p, df1, df2)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{df1}{degrees of freedom for the numerator (df1 > 0)}

\item{df2}{degrees of freedom for the denominator (df2 > 0)}
}
\value{
Quantile function value at \code{p}
}
\description{
Fisher F Distribution Quantile Function
}
\examples{
fisher_f_quantile(0.5, 5, 2)
}
