% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_distribution.R
\name{gamma_cdf}
\alias{gamma_cdf}
\title{Gamma Cumulative Distribution Function}
\usage{
gamma_cdf(x, shape, scale)
}
\arguments{
\item{x}{quantile}

\item{shape}{shape parameter (shape > 0)}

\item{scale}{scale parameter (scale > 0)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Gamma Cumulative Distribution Function
}
\examples{
gamma_cdf(2, 3, 4)
}
