% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_distribution.R
\name{gamma_lcdf}
\alias{gamma_lcdf}
\title{Gamma Log Cumulative Distribution Function}
\usage{
gamma_lcdf(x, shape, scale)
}
\arguments{
\item{x}{quantile}

\item{shape}{shape parameter (shape > 0)}

\item{scale}{scale parameter (scale > 0)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Gamma Log Cumulative Distribution Function
}
\examples{
gamma_lcdf(2, 3, 4)
}
