% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_distribution.R
\name{gamma_lpdf}
\alias{gamma_lpdf}
\title{Gamma Log Probability Density Function}
\usage{
gamma_lpdf(x, shape, scale)
}
\arguments{
\item{x}{quantile}

\item{shape}{shape parameter (shape > 0)}

\item{scale}{scale parameter (scale > 0)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Gamma Log Probability Density Function
}
\examples{
gamma_lpdf(2, 3, 4)
}
