% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{gamma_p}
\alias{gamma_p}
\title{Normalised Lower Incomplete Gamma Function}
\usage{
gamma_p(a, z)
}
\arguments{
\item{a}{Shape parameter of the gamma function}

\item{z}{Upper limit of integration}
}
\value{
The value of the normalised lower incomplete gamma function P(a, z).
}
\description{
Computes the normalised lower incomplete gamma function P(a, z) = gamma(a, z) / gamma(a).
}
