% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{gamma_p_derivative}
\alias{gamma_p_derivative}
\title{Derivative of the Normalised Upper Incomplete Gamma Function}
\usage{
gamma_p_derivative(a, z)
}
\arguments{
\item{a}{Shape parameter of the gamma function}

\item{z}{Upper limit of integration}
}
\value{
The value of the derivative of the normalised upper incomplete gamma function Q(a, z).
}
\description{
Computes the derivative of the normalised upper incomplete gamma function Q(a, z) with respect to z.
}
