% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{gamma_p_inv}
\alias{gamma_p_inv}
\title{Inverse of the Normalised Lower Incomplete Gamma Function}
\usage{
gamma_p_inv(a, p)
}
\arguments{
\item{a}{Shape parameter of the gamma function}

\item{p}{Probability value}
}
\value{
The value of z such that P(a, z) = p.
}
\description{
Computes the inverse of the normalised lower incomplete gamma function P(a, z) = p.
}
