% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{gamma_p_inva}
\alias{gamma_p_inva}
\title{Inverse of the Normalised Lower Incomplete Gamma Function (with respect to a)}
\usage{
gamma_p_inva(z, p)
}
\arguments{
\item{z}{Upper limit of integration}

\item{p}{Probability value}
}
\value{
The value of a such that P(a, z) = p.
}
\description{
Computes the inverse of the normalised lower incomplete gamma function P(a, z) = p,
with respect to the shape parameter a.
}
