% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_distribution.R
\name{gamma_pdf}
\alias{gamma_pdf}
\title{Gamma Probability Density Function}
\usage{
gamma_pdf(x, shape, scale)
}
\arguments{
\item{x}{quantile}

\item{shape}{shape parameter (shape > 0)}

\item{scale}{scale parameter (scale > 0)}
}
\value{
Probability density function value at \code{x}
}
\description{
Gamma Probability Density Function
}
\examples{
gamma_pdf(2, 3, 4)
}
