% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{gamma_q}
\alias{gamma_q}
\title{Normalised Upper Incomplete Gamma Function}
\usage{
gamma_q(a, z)
}
\arguments{
\item{a}{Shape parameter of the gamma function}

\item{z}{Upper limit of integration}
}
\value{
The value of the normalised upper incomplete gamma function Q(a, z).
}
\description{
Computes the normalised upper incomplete gamma function Q(a, z) = gamma(a, z) / gamma(a).
}
