% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{gamma_q_inv}
\alias{gamma_q_inv}
\title{Inverse of the Normalised Upper Incomplete Gamma Function}
\usage{
gamma_q_inv(a, q)
}
\arguments{
\item{a}{Shape parameter of the gamma function}

\item{q}{Probability value}
}
\value{
The value of z such that Q(a, z) = q.
}
\description{
Computes the inverse of the normalised upper incomplete gamma function Q(a, z) = q.
}
