% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{gamma_q_inva}
\alias{gamma_q_inva}
\title{Inverse of the Normalised Upper Incomplete Gamma Function (with respect to a)}
\usage{
gamma_q_inva(z, q)
}
\arguments{
\item{z}{Upper limit of integration}

\item{q}{Probability value}
}
\value{
The value of a such that Q(a, z) = q.
}
\description{
Computes the inverse of the normalised upper incomplete gamma function Q(a, z) = q,
with respect to the shape parameter a.
}
