% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_distribution.R
\name{gamma_quantile}
\alias{gamma_quantile}
\title{Gamma Quantile Function}
\usage{
gamma_quantile(p, shape, scale)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{shape}{shape parameter (shape > 0)}

\item{scale}{scale parameter (scale > 0)}
}
\value{
Quantile function value at \code{p}
}
\description{
Gamma Quantile Function
}
\examples{
gamma_quantile(0.5, 3, 4)
}
