% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature_and_differentiation.R
\name{gauss_kronrod}
\alias{gauss_kronrod}
\title{Gauss-Kronrod Quadrature for Numerical Integration}
\usage{
gauss_kronrod(f, a, b, points = 5)
}
\arguments{
\item{f}{A function to integrate. It should accept a single numeric value and
return a single numeric value.}

\item{a}{The lower limit of integration.}

\item{b}{The upper limit of integration.}

\item{points}{The number of evaluation points to use in the Gauss-Kronrod
quadrature. Must be an odd number between 1 and 100.}
}
\value{
The approximate value of the integral.
}
\description{
Performs adaptive Gauss-Kronrod integration using a specified number of evaluation points.
}
