% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature.R
\name{gauss_legendre}
\alias{gauss_legendre}
\title{Gauss-Legendre Quadrature for Numerical Integration}
\usage{
gauss_legendre(f, a, b, points = 7)
}
\arguments{
\item{f}{A function to integrate. It should accept a single numeric value and
return a single numeric value.}

\item{a}{The lower limit of integration.}

\item{b}{The upper limit of integration.}

\item{points}{The number of evaluation points to use in the Gauss-Legendre
quadrature. Default is 7.}
}
\value{
The approximate value of the integral.
}
\description{
This function performs non-adaptive Gauss-Legendre integration using a
specified number of evaluation points.
}
