% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{gegenbauer}
\alias{gegenbauer}
\title{Gegenbauer Polynomial}
\usage{
gegenbauer(n, lambda, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{lambda}{Parameter of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the Gegenbauer polynomial \eqn{\(C_n^{(\lambda)}(x)\)}.
}
\description{
Computes the Gegenbauer polynomial \eqn{\(C_n^{(\lambda)}(x)\)}.
}
