% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{gegenbauer_derivative}
\alias{gegenbauer_derivative}
\title{Gegenbauer Polynomial (Derivative with Order k)}
\usage{
gegenbauer_derivative(n, lambda, x, k)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{lambda}{Parameter of the polynomial}

\item{x}{Argument of the polynomial}

\item{k}{Order of the derivative}
}
\value{
The value of the k-th derivative of the Gegenbauer polynomial \eqn{\(C_n^{(\lambda)}(x)\)}.
}
\description{
Computes the k-th derivative of the Gegenbauer polynomial \eqn{\(C_n^{(\lambda)}(x)\)}.
}
