% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_distribution.R
\name{geometric_cdf}
\alias{geometric_cdf}
\title{Geometric Cumulative Distribution Function}
\usage{
geometric_cdf(x, prob)
}
\arguments{
\item{x}{quantile (non-negative integer)}

\item{prob}{probability of success (0 < prob < 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Geometric Cumulative Distribution Function
}
\examples{
geometric_cdf(3, 0.5)
}
