% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_distribution.R
\name{geometric_lcdf}
\alias{geometric_lcdf}
\title{Geometric Log Cumulative Distribution Function}
\usage{
geometric_lcdf(x, prob)
}
\arguments{
\item{x}{quantile (non-negative integer)}

\item{prob}{probability of success (0 < prob < 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Geometric Log Cumulative Distribution Function
}
\examples{
geometric_lcdf(3, 0.5)
}
