% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_distribution.R
\name{geometric_lpdf}
\alias{geometric_lpdf}
\title{Geometric Log Probability Density Function}
\usage{
geometric_lpdf(x, prob)
}
\arguments{
\item{x}{quantile (non-negative integer)}

\item{prob}{probability of success (0 < prob < 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Geometric Log Probability Density Function
}
\examples{
geometric_lpdf(3, 0.5)
}
