% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_distribution.R
\name{geometric_pdf}
\alias{geometric_pdf}
\title{Geometric Probability Density Function}
\usage{
geometric_pdf(x, prob)
}
\arguments{
\item{x}{quantile (non-negative integer)}

\item{prob}{probability of success (0 < prob < 1)}
}
\value{
Probability density function value at \code{x}
}
\description{
Geometric Probability Density Function
}
\examples{
geometric_pdf(3, 0.5)
}
