% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_distribution.R
\name{geometric_quantile}
\alias{geometric_quantile}
\title{Geometric Quantile Function}
\usage{
geometric_quantile(p, prob)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{prob}{probability of success (0 < prob < 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Geometric Quantile Function
}
\examples{
geometric_quantile(0.5, 0.5)
}
