% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootfinding_and_minimisation.R
\name{halley_iterate}
\alias{halley_iterate}
\title{Root-Finding with Halley's Method}
\usage{
halley_iterate(
  f,
  guess,
  min,
  max,
  digits = .Machine$double.digits,
  max_iter = .Machine$integer.max
)
}
\arguments{
\item{f}{A function to find the root of. It should take a single numeric value and return a three-element numeric vector: the function value, its first derivative, and its second derivative.}

\item{guess}{A numeric value that is a guess for the root.}

\item{min}{The minimum value of the root.}

\item{max}{The maximum value of the root.}

\item{digits}{The number of significant digits to which the root should be found. Defaults to double precision.}

\item{max_iter}{The maximum number of iterations to perform. Defaults to the maximum integer value.}
}
\value{
The function root, with attribute \code{iterations} indicating the number of iterations performed.
}
\description{
Root-Finding with Halley's Method
}
\examples{
f <- function(x) c(x^2 - 2, 2 * x, 2)
halley_iterate(f, guess = 1, min = 0, max = 2)
}
