% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_functionals.R
\name{hamming_distance}
\alias{hamming_distance}
\title{Hamming Distance}
\usage{
hamming_distance(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}
}
\value{
The Hamming distance between \code{x} and \code{y}.
}
\description{
Computes the Hamming distance between two vectors.
}
\examples{
hamming_distance(c(1, 0, 1), c(0, 1, 1))
}
