% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hankel_functions.R
\name{hankel_functions}
\alias{hankel_functions}
\alias{cyl_hankel_1}
\alias{cyl_hankel_2}
\alias{sph_hankel_1}
\alias{sph_hankel_2}
\title{Hankel Functions}
\usage{
cyl_hankel_1(v, x)

cyl_hankel_2(v, x)

sph_hankel_1(v, x)

sph_hankel_2(v, x)
}
\arguments{
\item{v}{Order of the Hankel function (can be any real number)}

\item{x}{Argument of the Hankel function (can be any real number)}
}
\value{
A single complex value with the computed Hankel function.
}
\description{
Functions to compute cylindrical and spherical Hankel functions of the first and second kinds.

\strong{Cyclic Hankel Functions}
\itemize{
\item \code{cyl_hankel_1(v, x)}: The Hankel function of the first kind: \eqn{H_v^{(1)}(x) = J_v(x) + iY_v(x)}
\item \code{cyl_hankel_2(v, x)}: The Hankel function of the second kind: \eqn{H_v^{(2)}(x) = J_v(x) - iY_v(x)}
}

Where \eqn{J_v(x)} is the Bessel function of the first kind and \eqn{Y_v(x)} is the Bessel function
of the second kind.

\strong{Spherical Hankel Functions:}
\itemize{
\item \code{sph_hankel_1(v, x)}: The spherical Hankel function of the first kind: \eqn{h_v^{(1)}(x)  = \sqrt{\frac{\pi}{2}}\frac{1}{\sqrt{\pi}}H_{v + \frac{1}{2}}^{(1)}(x)}
\item \code{sph_hankel_2(v, x)}: The spherical Hankel function of the second kind: \eqn{h_v^{(2)}(x)  = \sqrt{\frac{\pi}{2}}\frac{1}{\sqrt{\pi}}H_{v + \frac{1}{2}}^{(2)}(x)}
}
}
\examples{
cyl_hankel_1(2, 0.5)
cyl_hankel_2(2, 0.5)
sph_hankel_1(2, 0.5)
sph_hankel_2(2, 0.5)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/hankel.html}{Boost Documentation} for more details on the mathematical background.
}
