% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{hermite}
\alias{hermite}
\title{Hermite Polynomial}
\usage{
hermite(n, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the Hermite polynomial \eqn{\(H_n(x)\)}.
}
\description{
Computes the Hermite polynomial \eqn{\(H_n(x)\)}.
}
