% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{hermite_next}
\alias{hermite_next}
\title{Next Hermite Polynomial}
\usage{
hermite_next(n, x, Hn, Hnm1)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{x}{Argument of the polynomial}

\item{Hn}{Value of the Hermite polynomial \(H_n(x)\)}

\item{Hnm1}{Value of the Hermite polynomial \(H_{n-1}(x)\)}
}
\value{
The value of the next Hermite polynomial \(H_{n+1}(x)\).
}
\description{
Computes the next Hermite polynomial \(H_{n+1}(x)\) using the previous two polynomials.
}
