% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_integrals.R
\name{heuman_lambda}
\alias{heuman_lambda}
\title{Heuman's Lambda Function}
\usage{
heuman_lambda(k, phi)
}
\arguments{
\item{k}{Elliptic modulus}

\item{phi}{Amplitude}
}
\value{
The value of Heuman's lambda function.
}
\description{
Computes Heuman's lambda function.
}
