% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperexponential_distribution.R
\name{hyperexponential_cdf}
\alias{hyperexponential_cdf}
\title{Hyperexponential Cumulative Distribution Function}
\usage{
hyperexponential_cdf(x, probabilities, rates)
}
\arguments{
\item{x}{quantile}

\item{probabilities}{vector of probabilities (sum must be 1)}

\item{rates}{vector of rates (all rates must be > 0)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Hyperexponential Cumulative Distribution Function
}
\examples{
hyperexponential_cdf(2, c(0.5, 0.5), c(1, 2))
}
