% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperexponential_distribution.R
\name{hyperexponential_lpdf}
\alias{hyperexponential_lpdf}
\title{Hyperexponential Log Probability Density Function}
\usage{
hyperexponential_lpdf(x, probabilities, rates)
}
\arguments{
\item{x}{quantile}

\item{probabilities}{vector of probabilities (sum must be 1)}

\item{rates}{vector of rates (all rates must be > 0)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Hyperexponential Log Probability Density Function
}
\examples{
hyperexponential_lpdf(2, c(0.5, 0.5), c(1, 2))
}
