% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperexponential_distribution.R
\name{hyperexponential_quantile}
\alias{hyperexponential_quantile}
\title{Hyperexponential Quantile Function}
\usage{
hyperexponential_quantile(p, probabilities, rates)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{probabilities}{vector of probabilities (sum must be 1)}

\item{rates}{vector of rates (all rates must be > 0)}
}
\value{
Quantile function value at \code{p}
}
\description{
Hyperexponential Quantile Function
}
\examples{
hyperexponential_quantile(0.5, c(0.5, 0.5), c(1, 2))
}
