% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_functions.R
\name{hypergeometric_0F1}
\alias{hypergeometric_0F1}
\title{Hypergeometric Function 0F1}
\usage{
hypergeometric_0F1(b, z)
}
\arguments{
\item{b}{Parameter of the hypergeometric function}

\item{z}{Argument of the hypergeometric function}
}
\value{
The value of the hypergeometric function 0F1(b; z).
}
\description{
Computes the hypergeometric function 0F1(b; z).
}
