% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_functions.R
\name{hypergeometric_1F0}
\alias{hypergeometric_1F0}
\title{Hypergeometric Function 1F0}
\usage{
hypergeometric_1F0(a, z)
}
\arguments{
\item{a}{Parameter of the hypergeometric function}

\item{z}{Argument of the hypergeometric function}
}
\value{
The value of the hypergeometric function 1F0(a; z).
}
\description{
Computes the hypergeometric function 1F0(a; z).
}
