% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_functions.R
\name{hypergeometric_1F1_regularized}
\alias{hypergeometric_1F1_regularized}
\title{Regularised Hypergeometric Function 1F1}
\usage{
hypergeometric_1F1_regularized(a, b, z)
}
\arguments{
\item{a}{First parameter of the hypergeometric function}

\item{b}{Second parameter of the hypergeometric function}

\item{z}{Argument of the hypergeometric function}
}
\value{
The value of the regularised hypergeometric function 1F1(a, b; z).
}
\description{
Computes the regularised hypergeometric function 1F1(a, b; z) = hypergeometric_1F1(a, b; z) / gamma(b).
}
