% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_functions.R
\name{hypergeometric_2F0}
\alias{hypergeometric_2F0}
\title{Hypergeometric Function 2F0}
\usage{
hypergeometric_2F0(a1, a2, z)
}
\arguments{
\item{a1}{First parameter of the hypergeometric function}

\item{a2}{Second parameter of the hypergeometric function}

\item{z}{Argument of the hypergeometric function}
}
\value{
The value of the hypergeometric function 2F0(a1, a2; z).
}
\description{
Computes the hypergeometric function 2F0(a1, a2; z).
}
