% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_functions.R
\name{hypergeometric_functions}
\alias{hypergeometric_functions}
\alias{hypergeometric_1F0}
\alias{hypergeometric_0F1}
\alias{hypergeometric_2F0}
\alias{hypergeometric_1F1}
\alias{hypergeometric_1F1_regularized}
\alias{log_hypergeometric_1F1}
\alias{hypergeometric_pFq}
\title{Hypergeometric Functions}
\usage{
hypergeometric_1F0(a, z)

hypergeometric_0F1(b, z)

hypergeometric_2F0(a1, a2, z)

hypergeometric_1F1(a, b, z)

hypergeometric_1F1_regularized(a, b, z)

log_hypergeometric_1F1(a, b, z)

hypergeometric_pFq(a, b, z)
}
\arguments{
\item{a}{Parameter of the hypergeometric function (numerator parameter)}

\item{z}{Argument of the hypergeometric function}

\item{b}{Denominator parameter of the hypergeometric function}

\item{a1}{First numerator parameter (for 2F0)}

\item{a2}{Second numerator parameter (for 2F0)}
}
\value{
A single numeric value with the computed hypergeometric function.
}
\description{
Functions to compute various hypergeometric functions, which are solutions to hypergeometric
differential equations and generalize many special functions.

\strong{Specialised Hypergeometric Functions:}
\itemize{
\item \strong{hypergeometric_1F0(a, z):} : \deqn{_1F_0(a;z) = \sum_{n=0}^\infty \frac{(a)_nz^n}{n!}}
\item \strong{hypergeometric_0F1(b, z):} : \deqn{_0F_1(;b;z) = \sum_{n=0}^\infty \frac{z^n}{(b)_n!}}
\item \strong{hypergeometric_2F0(a1, a2, z):} : \deqn{_2F_0(a_1, a_2;z) = \sum_{n=0}^\infty \frac{(a_1)_n(a_2)_nz^n}{n!}}
\item \strong{hypergeometric_1F1(a, b, z):} : \deqn{_1F_1(a, b;z) = \sum_{n=0}^\infty \frac{(a)_nz^n}{(b)_n!}}
\item \strong{hypergeometric_1F1_regularized(a, b, z):} : \deqn{_1\tilde{F}_1(a, b;z) = \frac{_1F_1(a, b;z)}{\Gamma(b)}}
\item \strong{log_hypergeometric_1F1(a, b, z):} : Numerically stable implementation of \eqn{\ln _1F_1(a, b;z)}
}

\strong{Generalised Hypergeometric Function pFq:}
\itemize{
\item \code{hypergeometric_pFq(a, b, z)}: General form with p numerator parameters and
q denominator parameters:
}

\deqn{_pF_q\left(\{a_1,...,a_p\}, \{b_1,...,b_q\}; z\right) = \sum_{n=0}^{\infty} \frac{\prod_{j=1}^{p}(a_j)_n}{\prod_{j=1}^{q}(b_j)_n}\frac{z^n}{n!}}
}
\examples{
# Hypergeometric Function 1F0
hypergeometric_1F0(2, 0.2)
# Hypergeometric Function 0F1
hypergeometric_0F1(1, 0.8)
# Hypergeometric Function 2F0
hypergeometric_2F0(0.1, -1, 0.1)
# Hypergeometric Function 1F1 (Kummer's function)
hypergeometric_1F1(2, 3, 1)
# Regularised Hypergeometric Function 1F1
hypergeometric_1F1_regularized(2, 3, 1)
# Logarithm of the Hypergeometric Function 1F1
log_hypergeometric_1F1(2, 3, 1)
# Generalised Hypergeometric Function pFq (3F4 example)
hypergeometric_pFq(c(2, 3, 4), c(5, 6, 7, 8), 0.5)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/hypergeometric.html}{Boost Documentation} for more details on the mathematical background.
}
