% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_distribution.R
\name{hypergeometric_lcdf}
\alias{hypergeometric_lcdf}
\title{Hypergeometric Log Cumulative Distribution Function}
\usage{
hypergeometric_lcdf(x, r, n, N)
}
\arguments{
\item{x}{quantile}

\item{r}{number of successes in the population (r >= 0)}

\item{n}{number of draws (n >= 0)}

\item{N}{population size (N >= r)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Hypergeometric Log Cumulative Distribution Function
}
\examples{
hypergeometric_lcdf(3, 5, 10, 20)
}
