% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_distribution.R
\name{hypergeometric_lpdf}
\alias{hypergeometric_lpdf}
\title{Hypergeometric Log Probability Density Function}
\usage{
hypergeometric_lpdf(x, r, n, N)
}
\arguments{
\item{x}{quantile}

\item{r}{number of successes in the population (r >= 0)}

\item{n}{number of draws (n >= 0)}

\item{N}{population size (N >= r)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Hypergeometric Log Probability Density Function
}
\examples{
hypergeometric_lpdf(3, 5, 10, 20)
}
