% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_distribution.R
\name{hypergeometric_pdf}
\alias{hypergeometric_pdf}
\title{Hypergeometric Probability Density Function}
\usage{
hypergeometric_pdf(x, r, n, N)
}
\arguments{
\item{x}{quantile}

\item{r}{number of successes in the population (r >= 0)}

\item{n}{number of draws (n >= 0)}

\item{N}{population size (N >= r)}
}
\value{
Probability density function value at \code{x}
}
\description{
Hypergeometric Probability Density Function
}
\examples{
hypergeometric_pdf(3, 5, 10, 20)
}
