% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_functions.R
\name{hypergeometric_pFq}
\alias{hypergeometric_pFq}
\title{Hypergeometric Function pFq}
\usage{
hypergeometric_pFq(a, b, z)
}
\arguments{
\item{a}{Vector of parameters for the numerator of the hypergeometric function}

\item{b}{Vector of parameters for the denominator of the hypergeometric function}

\item{z}{Argument of the hypergeometric function}
}
\value{
The value of the hypergeometric function pFq(a, b; z).
}
\description{
Computes the generalised hypergeometric function pFq(a, b; z).
}
