% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_distribution.R
\name{hypergeometric_quantile}
\alias{hypergeometric_quantile}
\title{Hypergeometric Quantile Function}
\usage{
hypergeometric_quantile(p, r, n, N)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{r}{number of successes in the population (r >= 0)}

\item{n}{number of draws (n >= 0)}

\item{N}{population size (N >= r)}
}
\value{
Quantile function value at \code{p}
}
\description{
Hypergeometric Quantile Function
}
\examples{
hypergeometric_quantile(0.5, 5, 10, 20)
}
