% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_functions.R
\name{ibeta}
\alias{ibeta}
\title{Normalised Incomplete Beta Function}
\usage{
ibeta(a, b, x)
}
\arguments{
\item{a}{First parameter of the beta function}

\item{b}{Second parameter of the beta function}

\item{x}{Upper limit of integration}
}
\value{
The value of the normalised incomplete beta function I_x(a, b).
}
\description{
Calculates the normalised incomplete beta function I_x(a, b).
}
