% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_functions.R
\name{ibeta_derivative}
\alias{ibeta_derivative}
\title{Derivative of the Incomplete Beta Function}
\usage{
ibeta_derivative(a, b, x)
}
\arguments{
\item{a}{First parameter of the beta function}

\item{b}{Second parameter of the beta function}

\item{x}{Upper limit of integration}
}
\value{
The value of the derivative of the incomplete beta function with respect to x.
}
\description{
Calculates the derivative of the incomplete beta function with respect to x.
}
