% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_functions.R
\name{ibeta_inv}
\alias{ibeta_inv}
\title{Inverse of the Normalised Incomplete Beta Function}
\usage{
ibeta_inv(a, b, p)
}
\arguments{
\item{a}{First parameter of the beta function}

\item{b}{Second parameter of the beta function}

\item{p}{Probability value}
}
\value{
The value of x such that I_x(a, b) = p.
}
\description{
Calculates the inverse of the normalised incomplete beta function I_x(a, b) = p.
}
