% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_functions.R
\name{ibeta_inva}
\alias{ibeta_inva}
\title{Inverse of the Normalised Incomplete Beta Function with respect to a}
\usage{
ibeta_inva(b, x, p)
}
\arguments{
\item{b}{Second parameter of the beta function}

\item{x}{Upper limit of integration}

\item{p}{Probability value}
}
\value{
The value of a such that I_x(a, b) = p.
}
\description{
Calculates the inverse of the normalised incomplete beta function I_x(a, b) = p with respect to a.
}
