% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_functions.R
\name{ibeta_invb}
\alias{ibeta_invb}
\title{Inverse of the Normalised Incomplete Beta Function with respect to b}
\usage{
ibeta_invb(a, x, p)
}
\arguments{
\item{a}{First parameter of the beta function}

\item{x}{Upper limit of integration}

\item{p}{Probability value}
}
\value{
The value of b such that I_x(a, b) = p.
}
\description{
Calculates the inverse of the normalised incomplete beta function I_x(a, b) = p with respect to b.
}
