% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_functions.R
\name{ibetac}
\alias{ibetac}
\title{Normalised Complement of the Incomplete Beta Function}
\usage{
ibetac(a, b, x)
}
\arguments{
\item{a}{First parameter of the beta function}

\item{b}{Second parameter of the beta function}

\item{x}{Upper limit of integration}
}
\value{
The value of the normalised complement of the incomplete beta function 1 - I_x(a, b).
}
\description{
Calculates the normalised complement of the incomplete beta function 1 - I_x(a, b).
}
