% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_functions.R
\name{ibetac_inv}
\alias{ibetac_inv}
\title{Inverse of the Normalised Complement of the Incomplete Beta Function}
\usage{
ibetac_inv(a, b, q)
}
\arguments{
\item{a}{First parameter of the beta function}

\item{b}{Second parameter of the beta function}

\item{q}{Probability value}
}
\value{
The value of x such that I_x(a, b) = q.
}
\description{
Calculates the inverse of the normalised complement of the incomplete beta function I_x(a, b) = q.
}
