% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_functions.R
\name{ibetac_inva}
\alias{ibetac_inva}
\title{Inverse of the Normalised Complement of the Incomplete Beta Function with respect to a}
\usage{
ibetac_inva(b, x, q)
}
\arguments{
\item{b}{Second parameter of the beta function}

\item{x}{Upper limit of integration}

\item{q}{Probability value}
}
\value{
The value of a such that I_x(a, b) = q.
}
\description{
Calculates the inverse of the normalised complement of the incomplete beta function I_x(a, b) = q with respect to a.
}
