% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_chi_squared_distribution.R
\name{inverse_chi_squared_lcdf}
\alias{inverse_chi_squared_lcdf}
\title{Inverse Chi-Squared Log Cumulative Distribution Function}
\usage{
inverse_chi_squared_lcdf(x, df, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (df > 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Inverse Chi-Squared Log Cumulative Distribution Function
}
\examples{
inverse_chi_squared_lcdf(2, 3, 1)
}
