% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_chi_squared_distribution.R
\name{inverse_chi_squared_lpdf}
\alias{inverse_chi_squared_lpdf}
\title{Inverse Chi-Squared Log Probability Density Function}
\usage{
inverse_chi_squared_lpdf(x, df, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (df > 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Inverse Chi-Squared Log Probability Density Function
}
\examples{
inverse_chi_squared_lpdf(2, 3, 1)
}
