% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_chi_squared_distribution.R
\name{inverse_chi_squared_quantile}
\alias{inverse_chi_squared_quantile}
\title{Inverse Chi-Squared Quantile Function}
\usage{
inverse_chi_squared_quantile(p, df, scale = 1)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{df}{degrees of freedom (df > 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Inverse Chi-Squared Quantile Function
}
\examples{
inverse_chi_squared_quantile(0.5, 3, 1)
}
