% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_gamma_distribution.R
\name{inverse_gamma_cdf}
\alias{inverse_gamma_cdf}
\title{Inverse Gamma Cumulative Distribution Function}
\usage{
inverse_gamma_cdf(x, shape, scale)
}
\arguments{
\item{x}{quantile}

\item{shape}{shape parameter (shape > 0)}

\item{scale}{scale parameter (scale > 0)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Inverse Gamma Cumulative Distribution Function
}
\examples{
inverse_gamma_cdf(2, 3, 4)
}
