% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_gamma_distribution.R
\name{inverse_gamma_pdf}
\alias{inverse_gamma_pdf}
\title{Inverse Gamma Probability Density Function}
\usage{
inverse_gamma_pdf(x, shape, scale)
}
\arguments{
\item{x}{quantile}

\item{shape}{shape parameter (shape > 0)}

\item{scale}{scale parameter (scale > 0)}
}
\value{
Probability density function value at \code{x}
}
\description{
Inverse Gamma Probability Density Function
}
\examples{
inverse_gamma_pdf(2, 3, 4)
}
