% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_gamma_distribution.R
\name{inverse_gamma_quantile}
\alias{inverse_gamma_quantile}
\title{Inverse Gamma Quantile Function}
\usage{
inverse_gamma_quantile(p, shape, scale)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{shape}{shape parameter (shape > 0)}

\item{scale}{scale parameter (scale > 0)}
}
\value{
Quantile function value at \code{p}
}
\description{
Inverse Gamma Quantile Function
}
\examples{
inverse_gamma_quantile(0.5, 3, 4)
}
