% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_gaussian_distribution.R
\name{inverse_gaussian_cdf}
\alias{inverse_gaussian_cdf}
\title{Inverse Gaussian Cumulative Distribution Function}
\usage{
inverse_gaussian_cdf(x, mu, lambda)
}
\arguments{
\item{x}{quantile}

\item{mu}{mean parameter (mu > 0)}

\item{lambda}{scale (precision) parameter (lambda > 0)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Inverse Gaussian Cumulative Distribution Function
}
\examples{
inverse_gaussian_cdf(2, 3, 4)
}
