% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_gaussian_distribution.R
\name{inverse_gaussian_pdf}
\alias{inverse_gaussian_pdf}
\title{Inverse Gaussian Probability Density Function}
\usage{
inverse_gaussian_pdf(x, mu, lambda)
}
\arguments{
\item{x}{quantile}

\item{mu}{mean parameter (mu > 0)}

\item{lambda}{scale (precision) parameter (lambda > 0)}
}
\value{
Probability density function value at \code{x}
}
\description{
Inverse Gaussian Probability Density Function
}
\examples{
inverse_gaussian_pdf(2, 3, 4)
}
