% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_gaussian_distribution.R
\name{inverse_gaussian_quantile}
\alias{inverse_gaussian_quantile}
\title{Inverse Gaussian Quantile Function}
\usage{
inverse_gaussian_quantile(p, mu, lambda)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{mu}{mean parameter (mu > 0)}

\item{lambda}{scale (precision) parameter (lambda > 0)}
}
\value{
Quantile function value at \code{p}
}
\description{
Inverse Gaussian Quantile Function
}
\examples{
inverse_gaussian_quantile(0.5, 3, 4)
}
