% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{jacobi}
\alias{jacobi}
\title{Jacobi Polynomial}
\usage{
jacobi(n, alpha, beta, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{alpha}{First parameter of the polynomial}

\item{beta}{Second parameter of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the Jacobi polynomial \eqn{\(P_n^{(\alpha, \beta)}(x)\)}.
}
\description{
Computes the Jacobi polynomial \eqn{\(P_n^{(\alpha, \beta)}(x)\)}.
}
