% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_cd}
\alias{jacobi_cd}
\title{Jacobi Elliptic CD Function}
\usage{
jacobi_cd(k, u)
}
\arguments{
\item{k}{Elliptic modulus}

\item{u}{Argument of the elliptic function}
}
\value{
The value of the Jacobi elliptic function cd.
}
\description{
Computes the Jacobi elliptic function cd.
}
