% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_cn}
\alias{jacobi_cn}
\title{Jacobi Elliptic CN Function}
\usage{
jacobi_cn(k, u)
}
\arguments{
\item{k}{Elliptic modulus}

\item{u}{Argument of the elliptic function}
}
\value{
The value of the Jacobi elliptic function cn.
}
\description{
Computes the Jacobi elliptic function cn.
}
