% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_cs}
\alias{jacobi_cs}
\title{Jacobi Elliptic CS Function}
\usage{
jacobi_cs(k, u)
}
\arguments{
\item{k}{Elliptic modulus}

\item{u}{Argument of the elliptic function}
}
\value{
The value of the Jacobi elliptic function cs.
}
\description{
Computes the Jacobi elliptic function cs.
}
