% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_dc}
\alias{jacobi_dc}
\title{Jacobi Elliptic DC Function}
\usage{
jacobi_dc(k, u)
}
\arguments{
\item{k}{Elliptic modulus}

\item{u}{Argument of the elliptic function}
}
\value{
The value of the Jacobi elliptic function dc.
}
\description{
Computes the Jacobi elliptic function dc.
}
