% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{jacobi_derivative}
\alias{jacobi_derivative}
\title{Jacobi Polynomial (Derivative with Order k)}
\usage{
jacobi_derivative(n, alpha, beta, x, k)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{alpha}{First parameter of the polynomial}

\item{beta}{Second parameter of the polynomial}

\item{x}{Argument of the polynomial}

\item{k}{Order of the derivative}
}
\value{
The value of the k-th derivative of the Jacobi polynomial \(P_n^{(\alpha, \beta)}(x)\).
}
\description{
Computes the k-th derivative of the Jacobi polynomial \(P_n^{(\alpha, \beta)}(x)\).
}
