% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_dn}
\alias{jacobi_dn}
\title{Jacobi Elliptic DN Function}
\usage{
jacobi_dn(k, u)
}
\arguments{
\item{k}{Elliptic modulus}

\item{u}{Argument of the elliptic function}
}
\value{
The value of the Jacobi elliptic function dn.
}
\description{
Computes the Jacobi elliptic function dn.
}
