% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{jacobi_double_prime}
\alias{jacobi_double_prime}
\title{Jacobi Polynomial (Second Derivative)}
\usage{
jacobi_double_prime(n, alpha, beta, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{alpha}{First parameter of the polynomial}

\item{beta}{Second parameter of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the second derivative of the Jacobi polynomial \(P_n^{(\alpha, \beta)}(x)\).
}
\description{
Computes the second derivative of the Jacobi polynomial \(P_n^{(\alpha, \beta)}(x)\).
}
