% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_ds}
\alias{jacobi_ds}
\title{Jacobi Elliptic DS Function}
\usage{
jacobi_ds(k, u)
}
\arguments{
\item{k}{Elliptic modulus}

\item{u}{Argument of the elliptic function}
}
\value{
The value of the Jacobi elliptic function ds.
}
\description{
Computes the Jacobi elliptic function ds.
}
