% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_elliptic}
\alias{jacobi_elliptic}
\title{Jacobi Elliptic SN, CN, and DN}
\usage{
jacobi_elliptic(k, u)
}
\arguments{
\item{k}{Elliptic modulus}

\item{u}{Argument of the elliptic functions}
}
\value{
A list containing the values of the Jacobi elliptic functions sn, cn, and dn.
}
\description{
Computes the Jacobi elliptic functions sn, cn, and dn.
}
