% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_elliptic_functions}
\alias{jacobi_elliptic_functions}
\alias{jacobi_elliptic}
\alias{jacobi_cd}
\alias{jacobi_cn}
\alias{jacobi_cs}
\alias{jacobi_dc}
\alias{jacobi_dn}
\alias{jacobi_ds}
\alias{jacobi_nc}
\alias{jacobi_nd}
\alias{jacobi_ns}
\alias{jacobi_sc}
\alias{jacobi_sd}
\alias{jacobi_sn}
\title{Jacobi Elliptic Functions}
\usage{
jacobi_elliptic(k, u)

jacobi_cd(k, u)

jacobi_cn(k, u)

jacobi_cs(k, u)

jacobi_dc(k, u)

jacobi_dn(k, u)

jacobi_ds(k, u)

jacobi_nc(k, u)

jacobi_nd(k, u)

jacobi_ns(k, u)

jacobi_sc(k, u)

jacobi_sd(k, u)

jacobi_sn(k, u)
}
\arguments{
\item{k}{Elliptic modulus (typically 0 <= k <= 1, but k > 1 uses complementary relations)}

\item{u}{Argument of the elliptic functions}
}
\value{
For \code{jacobi_elliptic}, a list containing the values of the Jacobi elliptic functions: \code{sn}, \code{cn}, \code{dn}. For individual functions, a single numeric value is returned.
}
\description{
Functions to compute the Jacobi elliptic functions, which are doubly periodic
generalizations of trigonometric and hyperbolic functions.

\strong{Jacobi Elliptic SN, CN and DN}
Given:

\deqn{u = \int_0^\phi\frac{d\psi}{\sqrt{1-k^2\text{sin}^2\psi}}}
\itemize{
\item \code{jacobi_sn(k, u)}: \eqn{\text{sin}\psi}
\item \code{jacobi_cn(k, u)}: \eqn{\text{cos}\psi}
\item \code{jacobi_dn(k, u)}: \eqn{\sqrt{(1 - k^2\text{sin}^2\psi)}}
\item \code{jacobi_cd(k, u)}: \eqn{\frac{cn(k, u)}{dn(k, u)}}
\item \code{jacobi_cs(k, u)}: \eqn{\frac{cn(k, u)}{sn(k, u)}}
\item \code{jacobi_dc(k, u)}: \eqn{\frac{dn(k, u)}{cn(k, u)}}
\item \code{jacobi_ds(k, u)}: \eqn{\frac{dn(k, u)}{sn(k, u)}}
\item \code{jacobi_nc(k, u)}: \eqn{\frac{1}{cn(k, u)}}
\item \code{jacobi_nd(k, u)}: \eqn{\frac{1}{dn(k, u)}}
\item \code{jacobi_ns(k, u)}: \eqn{\frac{1}{sn(k, u)}}
\item \code{jacobi_sc(k, u)}: \eqn{\frac{sn(k, u)}{cn(k, u)}}
\item \code{jacobi_sd(k, u)}: \eqn{\frac{sn(k, u)}{dn(k, u)}}
}
}
\examples{
# All three principal Jacobi Elliptic Functions at once
k <- 0.5
u <- 2
jacobi_elliptic(k, u)
# Individual Jacobi Elliptic Functions
jacobi_cd(k, u)
jacobi_cn(k, u)
jacobi_cs(k, u)
jacobi_dc(k, u)
jacobi_dn(k, u)
jacobi_ds(k, u)
jacobi_nc(k, u)
jacobi_nd(k, u)
jacobi_ns(k, u)
jacobi_sc(k, u)
jacobi_sd(k, u)
jacobi_sn(k, u)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/jacobi.html}{Boost Documentation} for more details on the mathematical background.
}
