% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_nc}
\alias{jacobi_nc}
\title{Jacobi Elliptic NC Function}
\usage{
jacobi_nc(k, u)
}
\arguments{
\item{k}{Elliptic modulus}

\item{u}{Argument of the elliptic function}
}
\value{
The value of the Jacobi elliptic function nc.
}
\description{
Computes the Jacobi elliptic function nc.
}
