% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_nd}
\alias{jacobi_nd}
\title{Jacobi Elliptic ND Function}
\usage{
jacobi_nd(k, u)
}
\arguments{
\item{k}{Elliptic modulus}

\item{u}{Argument of the elliptic function}
}
\value{
The value of the Jacobi elliptic function nd.
}
\description{
Computes the Jacobi elliptic function nd.
}
