% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_ns}
\alias{jacobi_ns}
\title{Jacobi Elliptic NS Function}
\usage{
jacobi_ns(k, u)
}
\arguments{
\item{k}{Elliptic modulus}

\item{u}{Argument of the elliptic function}
}
\value{
The value of the Jacobi elliptic function ns.
}
\description{
Computes the Jacobi elliptic function ns.
}
