% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{jacobi_prime}
\alias{jacobi_prime}
\title{Jacobi Polynomial (Derivative)}
\usage{
jacobi_prime(n, alpha, beta, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{alpha}{First parameter of the polynomial}

\item{beta}{Second parameter of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the derivative of the Jacobi polynomial \eqn{(P_n^{(\alpha, \beta)}(x))}.
}
\description{
Computes the derivative of the Jacobi polynomial \eqn{(P_n^{(\alpha, \beta)}(x))}.
}
