% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_sc}
\alias{jacobi_sc}
\title{Jacobi Elliptic SC Function}
\usage{
jacobi_sc(k, u)
}
\arguments{
\item{k}{Elliptic modulus}

\item{u}{Argument of the elliptic function}
}
\value{
The value of the Jacobi elliptic function sc.
}
\description{
Computes the Jacobi elliptic function sc.
}
