% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_sd}
\alias{jacobi_sd}
\title{Jacobi Elliptic SD Function}
\usage{
jacobi_sd(k, u)
}
\arguments{
\item{k}{Elliptic modulus}

\item{u}{Argument of the elliptic function}
}
\value{
The value of the Jacobi elliptic function sd.
}
\description{
Computes the Jacobi elliptic function sd.
}
