% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_elliptic_functions.R
\name{jacobi_sn}
\alias{jacobi_sn}
\title{Jacobi Elliptic SN Function}
\usage{
jacobi_sn(k, u)
}
\arguments{
\item{k}{Elliptic modulus}

\item{u}{Argument of the elliptic function}
}
\value{
The value of the Jacobi elliptic function sn.
}
\description{
Computes the Jacobi elliptic function sn.
}
