% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_theta_functions.R
\name{jacobi_theta2}
\alias{jacobi_theta2}
\title{Jacobi Theta2 Function (q parameterisation)}
\usage{
jacobi_theta2(x, q)
}
\arguments{
\item{x}{Input value}

\item{q}{The nome parameter of the Jacobi theta function}
}
\value{
The value of the Jacobi theta function \eqn{\(\theta_2(x; q)\)}.
}
\description{
Computes the Jacobi theta function \eqn{\(\theta_2\)} parameterised by \eqn{\(q\)}.
}
