% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_theta_functions.R
\name{jacobi_theta2tau}
\alias{jacobi_theta2tau}
\title{Jacobi Theta2 Function (tau parameterisation)}
\usage{
jacobi_theta2tau(x, tau)
}
\arguments{
\item{x}{Input value}

\item{tau}{The nome parameter of the Jacobi theta function}
}
\value{
The value of the Jacobi theta function \(\theta_2(x; \tau)\).
}
\description{
Computes the Jacobi theta function \(\theta_2\) parameterised by \(\tau\).
}
