% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_theta_functions.R
\name{jacobi_theta3}
\alias{jacobi_theta3}
\title{Jacobi Theta3 Function (q parameterisation)}
\usage{
jacobi_theta3(x, q)
}
\arguments{
\item{x}{Input value}

\item{q}{The nome parameter of the Jacobi theta function}
}
\value{
The value of the Jacobi theta function \(\theta_3(x; q)\).
}
\description{
Computes the Jacobi theta function \(\theta_3\) parameterised by \(q\).
}
