% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_theta_functions.R
\name{jacobi_theta4}
\alias{jacobi_theta4}
\title{Jacobi Theta4 Function (q parameterisation)}
\usage{
jacobi_theta4(x, q)
}
\arguments{
\item{x}{Input value}

\item{q}{The nome parameter of the Jacobi theta function}
}
\value{
The value of the Jacobi theta function \eqn{(\theta_4(x; q))}.
}
\description{
Computes the Jacobi theta function \eqn{(\theta_4)} parameterised by \eqn{(q)}.
}
