% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_theta_functions.R
\name{jacobi_theta4m1}
\alias{jacobi_theta4m1}
\title{Jacobi Theta4 Function (q parameterisation, minus one)}
\usage{
jacobi_theta4m1(x, q)
}
\arguments{
\item{x}{Input value}

\item{q}{The nome parameter of the Jacobi theta function}
}
\value{
The value of the Jacobi theta function \eqn{\(\theta_4(x; q) - 1\)}.
}
\description{
Computes the Jacobi theta function \eqn{\(\theta_4\)} parameterised by \eqn{\(q\)} with a shift of -1.
}
