% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobi_theta_functions.R
\name{jacobi_theta4tau}
\alias{jacobi_theta4tau}
\title{Jacobi Theta4 Function (tau parameterisation)}
\usage{
jacobi_theta4tau(x, tau)
}
\arguments{
\item{x}{Input value}

\item{tau}{The nome parameter of the Jacobi theta function}
}
\value{
The value of the Jacobi theta function \eqn{\(\theta_4(x; \tau)\)}.
}
\description{
Computes the Jacobi theta function \eqn{\(\theta_4\)$ parameterised by $\(\tau\)}.
}
