% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptic_integrals.R
\name{jacobi_zeta}
\alias{jacobi_zeta}
\title{Jacobi Zeta Function}
\usage{
jacobi_zeta(k, phi)
}
\arguments{
\item{k}{Elliptic modulus}

\item{phi}{Amplitude}
}
\value{
The value of the Jacobi zeta function.
}
\description{
Computes the Jacobi zeta function.
}
