% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolmogorov_smirnov_distribution.R
\name{kolmogorov_smirnov_cdf}
\alias{kolmogorov_smirnov_cdf}
\title{Kolmogorov-Smirnov Cumulative Distribution Function}
\usage{
kolmogorov_smirnov_cdf(x, n)
}
\arguments{
\item{x}{quantile}

\item{n}{sample size (n > 0)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Kolmogorov-Smirnov Cumulative Distribution Function
}
\examples{
kolmogorov_smirnov_cdf(0.5, 10)
}
