% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolmogorov_smirnov_distribution.R
\name{kolmogorov_smirnov_lcdf}
\alias{kolmogorov_smirnov_lcdf}
\title{Kolmogorov-Smirnov Log Cumulative Distribution Function}
\usage{
kolmogorov_smirnov_lcdf(x, n)
}
\arguments{
\item{x}{quantile}

\item{n}{sample size (n > 0)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Kolmogorov-Smirnov Log Cumulative Distribution Function
}
\examples{
kolmogorov_smirnov_lcdf(0.5, 10)
}
