% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolmogorov_smirnov_distribution.R
\name{kolmogorov_smirnov_lpdf}
\alias{kolmogorov_smirnov_lpdf}
\title{Kolmogorov-Smirnov Log Probability Density Function}
\usage{
kolmogorov_smirnov_lpdf(x, n)
}
\arguments{
\item{x}{quantile}

\item{n}{sample size (n > 0)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Kolmogorov-Smirnov Log Probability Density Function
}
\examples{
kolmogorov_smirnov_lpdf(0.5, 10)
}
