% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolmogorov_smirnov_distribution.R
\name{kolmogorov_smirnov_pdf}
\alias{kolmogorov_smirnov_pdf}
\title{Kolmogorov-Smirnov Probability Density Function}
\usage{
kolmogorov_smirnov_pdf(x, n)
}
\arguments{
\item{x}{quantile}

\item{n}{sample size (n > 0)}
}
\value{
Probability density function value at \code{x}
}
\description{
Kolmogorov-Smirnov Probability Density Function
}
\examples{
kolmogorov_smirnov_pdf(0.5, 10)
}
