% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolmogorov_smirnov_distribution.R
\name{kolmogorov_smirnov_quantile}
\alias{kolmogorov_smirnov_quantile}
\title{Kolmogorov-Smirnov Quantile Function}
\usage{
kolmogorov_smirnov_quantile(p, n)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{n}{sample size (n > 0)}
}
\value{
Quantile function value at \code{p}
}
\description{
Kolmogorov-Smirnov Quantile Function
}
\examples{
kolmogorov_smirnov_quantile(0.5, 10)
}
