% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_functionals.R
\name{l0_pseudo_norm}
\alias{l0_pseudo_norm}
\title{L0 Pseudo Norm}
\usage{
l0_pseudo_norm(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The L0 pseudo norm of the vector.
}
\description{
L0 Pseudo Norm
}
\examples{
l0_pseudo_norm(c(1, 0, 2, 0, 3))
}
