% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_functionals.R
\name{l1_distance}
\alias{l1_distance}
\title{L1 Distance}
\usage{
l1_distance(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}
}
\value{
The L1 distance between \code{x} and \code{y}.
}
\description{
Computes the L1 distance (Manhattan distance) between two vectors.
}
\examples{
l1_distance(c(1, -2, 3), c(4, -5, 6))
}
