% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_functionals.R
\name{l1_norm}
\alias{l1_norm}
\title{L1 Norm}
\usage{
l1_norm(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The L1 norm of the vector.
}
\description{
Computes the L1 norm (Manhattan norm) of a vector.
}
\examples{
l1_norm(c(1, -2, 3))
}
