% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_functionals.R
\name{l2_distance}
\alias{l2_distance}
\title{L2 Distance}
\usage{
l2_distance(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}
}
\value{
The L2 distance between \code{x} and \code{y}.
}
\description{
Computes the L2 distance (Euclidean distance) between two vectors.
}
\examples{
l2_distance(c(3, 4), c(0, 0))
}
