% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_functionals.R
\name{l2_norm}
\alias{l2_norm}
\title{L2 Norm}
\usage{
l2_norm(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The L2 norm of the vector.
}
\description{
Computes the L2 norm (Euclidean norm) of a vector.
}
\examples{
l2_norm(c(3, 4))
}
