% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{laguerre}
\alias{laguerre}
\title{Laguerre Polynomial}
\usage{
laguerre(n, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the Laguerre polynomial \(L_n(x)\).
}
\description{
Computes the Laguerre polynomial \(L_n(x)\).
}
