% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{laguerre_m}
\alias{laguerre_m}
\title{Laguerre Polynomial with Order m}
\usage{
laguerre_m(n, m, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{m}{Order of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the Laguerre polynomial \(L_n^m(x)\).
}
\description{
Computes the Laguerre polynomial \(L_n^m(x)\).
}
