% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{laguerre_next}
\alias{laguerre_next}
\title{Next Laguerre Polynomial}
\usage{
laguerre_next(n, x, Ln, Lnm1)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{x}{Argument of the polynomial}

\item{Ln}{Value of the Laguerre polynomial \eqn{(L_n(x))}}

\item{Lnm1}{Value of the Laguerre polynomial \eqn{(L_{n-1}(x))}}
}
\value{
The value of the next Laguerre polynomial \eqn{(L_{n+1}(x))}.
}
\description{
Computes the next Laguerre polynomial \eqn{(L_{n+1}(x))} using the previous two polynomials.
}
