% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{laguerre_next_m}
\alias{laguerre_next_m}
\title{Next Laguerre Polynomial with Order m}
\usage{
laguerre_next_m(n, m, x, Ln, Lnm1)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{m}{Order of the polynomial}

\item{x}{Argument of the polynomial}

\item{Ln}{Value of the Laguerre polynomial \eqn{\(L_n^m(x)\)}}

\item{Lnm1}{Value of the Laguerre polynomial \eqn{\(L_{n-1}^m(x)\)}}
}
\value{
The value of the next Laguerre polynomial \eqn{\(L_{n+1}^m(x)\)}.
}
\description{
Computes the next Laguerre polynomial \eqn{\(L_{n+1}^m(x)\)} using the previous two polynomials.
}
