% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambert_w_function.R
\name{lambert_w0}
\alias{lambert_w0}
\title{Lambert W Function (Principal Branch)}
\usage{
lambert_w0(z)
}
\arguments{
\item{z}{Argument of the Lambert W function}
}
\value{
The value of the Lambert W function \eqn{W_0(z)}.
}
\description{
Computes the Lambert W function for the principal branch (\eqn{W_0}).
}
