% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambert_w_function.R
\name{lambert_w0_prime}
\alias{lambert_w0_prime}
\title{Derivative of the Lambert W Function (Principal Branch)}
\usage{
lambert_w0_prime(z)
}
\arguments{
\item{z}{Argument of the Lambert W function}
}
\value{
The value of the derivative of the Lambert W function W_0'(z).
}
\description{
Computes the derivative of the Lambert W function for the principal branch (W_0').
}
