% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambert_w_function.R
\name{lambert_wm1}
\alias{lambert_wm1}
\title{Lambert W Function (Branch -1)}
\usage{
lambert_wm1(z)
}
\arguments{
\item{z}{Argument of the Lambert W function}
}
\value{
The value of the Lambert W function W_{-1}(z).
}
\description{
Computes the Lambert W function for the branch -1 (W_{-1}).
}
