% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambert_w_function.R
\name{lambert_wm1_prime}
\alias{lambert_wm1_prime}
\title{Derivative of the Lambert W Function (Branch -1)}
\usage{
lambert_wm1_prime(z)
}
\arguments{
\item{z}{Argument of the Lambert W function}
}
\value{
The value of the derivative of the Lambert W function \eqn{W_{-1}'}.
}
\description{
Computes the derivative of the Lambert W function for the branch -1 (\eqn{W_{-1}'}).
}
