% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landau_distribution.R
\name{landau_distribution}
\alias{landau_distribution}
\alias{landau_pdf}
\alias{landau_lpdf}
\alias{landau_cdf}
\alias{landau_lcdf}
\alias{landau_quantile}
\title{Landau Distribution Functions}
\usage{
landau_pdf(x, location = 0, scale = 1)

landau_lpdf(x, location = 0, scale = 1)

landau_cdf(x, location = 0, scale = 1)

landau_lcdf(x, location = 0, scale = 1)

landau_quantile(p, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}

\item{p}{probability (0 <= p <= 1)}
}
\description{
Functions to compute the probability density function, cumulative distribution function, and quantile function for the Landau distribution.
}
\examples{
# Distribution only available with Boost version 1.87.0 or later.
\dontrun{
# Landau distribution with location 0 and scale 1
  landau_pdf(3)
  landau_lpdf(3)
  landau_cdf(3)
  landau_lcdf(3)
  landau_quantile(0.5)
}
}
\seealso{
\href{https://www.boost.org/doc/libs/1_87_0/libs/math/doc/html/math_toolkit/dist_ref/dists/landau_dist.html}{Boost Documentation} for more details on the mathematical background.
}
