% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_distribution.R
\name{laplace_cdf}
\alias{laplace_cdf}
\title{Laplace Cumulative Distribution Function}
\usage{
laplace_cdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Laplace Cumulative Distribution Function
}
\examples{
laplace_cdf(0)
}
