% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_distribution.R
\name{laplace_lcdf}
\alias{laplace_lcdf}
\title{Laplace Log Cumulative Distribution Function}
\usage{
laplace_lcdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Laplace Log Cumulative Distribution Function
}
\examples{
laplace_lcdf(0)
}
