% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_distribution.R
\name{laplace_lpdf}
\alias{laplace_lpdf}
\title{Laplace Log Probability Density Function}
\usage{
laplace_lpdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Laplace Log Probability Density Function
}
\examples{
laplace_lpdf(0)
}
