% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_distribution.R
\name{laplace_pdf}
\alias{laplace_pdf}
\title{Laplace Probability Density Function}
\usage{
laplace_pdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Probability density function value at \code{x}
}
\description{
Laplace Probability Density Function
}
\examples{
laplace_pdf(0)
}
