% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_distribution.R
\name{laplace_quantile}
\alias{laplace_quantile}
\title{Laplace Quantile Function}
\usage{
laplace_quantile(p, location = 0, scale = 1)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Laplace Quantile Function
}
\examples{
laplace_quantile(0.5)
}
