% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{legendre_next}
\alias{legendre_next}
\title{Next Legendre Polynomial of the First Kind}
\usage{
legendre_next(l, x, Pl, Plm1)
}
\arguments{
\item{l}{Degree of the polynomial}

\item{x}{Argument of the polynomial}

\item{Pl}{Value of the Legendre polynomial \(P_l(x)\)}

\item{Plm1}{Value of the Legendre polynomial \(P_{l-1}(x)\)}
}
\value{
The value of the next Legendre polynomial \(P_{l+1}(x)\).
}
\description{
Computes the next Legendre polynomial of the first kind \(P_{l+1}(x)\) using the previous two polynomials.
}
