% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{legendre_next_m}
\alias{legendre_next_m}
\title{Next Legendre Polynomial of the First Kind with Order m}
\usage{
legendre_next_m(l, m, x, Pl, Plm1)
}
\arguments{
\item{l}{Degree of the polynomial}

\item{m}{Order of the polynomial}

\item{x}{Argument of the polynomial}

\item{Pl}{Value of the Legendre polynomial \eqn{\(P_l^m(x)\)}}

\item{Plm1}{Value of the Legendre polynomial \eqn{\(P_{l-1}^m(x)\)}}
}
\value{
The value of the next Legendre polynomial \eqn{\(P_{l+1}^m(x)\)}.
}
\description{
Computes the next Legendre polynomial of the first kind \eqn{\(P_{l+1}^m(x)\)} using the previous two polynomials.
}
