% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{legendre_p}
\alias{legendre_p}
\title{Legendre Polynomial of the First Kind}
\usage{
legendre_p(n, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the Legendre polynomial \eqn{\(P_n(x)\)}.
}
\description{
Computes the Legendre polynomial of the first kind \eqn{\(P_n(x)\)}.
}
