% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{legendre_p_m}
\alias{legendre_p_m}
\title{Legendre Polynomial of the First Kind with Order m}
\usage{
legendre_p_m(n, m, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{m}{Order of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the Legendre polynomial \eqn{(P_n^m(x))}.
}
\description{
Computes the Legendre polynomial of the first kind \eqn{(P_n^m(x))}.
}
