% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{legendre_p_prime}
\alias{legendre_p_prime}
\title{Legendre Polynomial of the First Kind (Derivative)}
\usage{
legendre_p_prime(n, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the derivative of the Legendre polynomial \eqn{\(P_n'(x)\)}.
}
\description{
Computes the derivative of the Legendre polynomial of the first kind \eqn{\(P_n'(x)\)}.
}
