% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{legendre_p_zeros}
\alias{legendre_p_zeros}
\title{Zeros of the Legendre Polynomial of the First Kind}
\usage{
legendre_p_zeros(l)
}
\arguments{
\item{l}{Degree of the polynomial}
}
\value{
A numeric vector containing the zeros of the Legendre polynomial \eqn{\(P_l(x)\)}.
}
\description{
Computes the zeros of the Legendre polynomial of the first kind \eqn{\(P_l(x)\)}.
}
