% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{legendre_q}
\alias{legendre_q}
\title{Legendre Polynomial of the Second Kind}
\usage{
legendre_q(n, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the Legendre polynomial \(Q_n(x)\).
}
\description{
Computes the Legendre polynomial of the second kind \(Q_n(x)\).
}
