% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{log1p_boost}
\alias{log1p_boost}
\title{Logarithm of 1 plus x}
\usage{
log1p_boost(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The natural logarithm of 1 plus x.
}
\description{
Computes the natural logarithm of 1 plus x.
}
