% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric_functions.R
\name{log_hypergeometric_1F1}
\alias{log_hypergeometric_1F1}
\title{Logarithm of the Hypergeometric Function 1F1}
\usage{
log_hypergeometric_1F1(a, b, z)
}
\arguments{
\item{a}{First parameter of the hypergeometric function}

\item{b}{Second parameter of the hypergeometric function}

\item{z}{Argument of the hypergeometric function}
}
\value{
The logarithm of the hypergeometric function 1F1(a, b; z) with attribute "sign" for the sign of the result.
}
\description{
Computes the logarithm of the hypergeometric function 1F1(a, b; z).
}
