% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_distribution.R
\name{logistic_cdf}
\alias{logistic_cdf}
\title{Logistic Cumulative Distribution Function}
\usage{
logistic_cdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Logistic Cumulative Distribution Function
}
\examples{
logistic_cdf(0)
}
