% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_distribution.R
\name{logistic_lcdf}
\alias{logistic_lcdf}
\title{Logistic Log Cumulative Distribution Function}
\usage{
logistic_lcdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Logistic Log Cumulative Distribution Function
}
\examples{
logistic_lcdf(0)
}
