% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_distribution.R
\name{logistic_lpdf}
\alias{logistic_lpdf}
\title{Logistic Log Probability Density Function}
\usage{
logistic_lpdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Logistic Log Probability Density Function
}
\examples{
logistic_lpdf(0)
}
