% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_distribution.R
\name{logistic_pdf}
\alias{logistic_pdf}
\title{Logistic Probability Density Function}
\usage{
logistic_pdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Probability density function value at \code{x}
}
\description{
Logistic Probability Density Function
}
\examples{
logistic_pdf(0)
}
