% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_distribution.R
\name{logistic_quantile}
\alias{logistic_quantile}
\title{Logistic Quantile Function}
\usage{
logistic_quantile(p, location = 0, scale = 1)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Logistic Quantile Function
}
\examples{
logistic_quantile(0.5)
}
