% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_distribution.R
\name{lognormal_cdf}
\alias{lognormal_cdf}
\title{Log Normal Cumulative Distribution Function}
\usage{
lognormal_cdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Log Normal Cumulative Distribution Function
}
\examples{
lognormal_cdf(0)
}
