% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_distribution.R
\name{lognormal_lcdf}
\alias{lognormal_lcdf}
\title{Log Normal Log Cumulative Distribution Function}
\usage{
lognormal_lcdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Log Normal Log Cumulative Distribution Function
}
\examples{
lognormal_lcdf(0)
}
