% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_distribution.R
\name{lognormal_lpdf}
\alias{lognormal_lpdf}
\title{Log Normal Log Probability Density Function}
\usage{
lognormal_lpdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Log Normal Log Probability Density Function
}
\examples{
lognormal_lpdf(0)
}
