% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_distribution.R
\name{lognormal_pdf}
\alias{lognormal_pdf}
\title{Log Normal Probability Density Function}
\usage{
lognormal_pdf(x, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Probability density function value at \code{x}
}
\description{
Log Normal Probability Density Function
}
\examples{
lognormal_pdf(0)
}
