% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_distribution.R
\name{lognormal_quantile}
\alias{lognormal_quantile}
\title{Log Normal Quantile Function}
\usage{
lognormal_quantile(p, location = 0, scale = 1)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Log Normal Quantile Function
}
\examples{
lognormal_quantile(0.5)
}
