% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_functionals.R
\name{lp_distance}
\alias{lp_distance}
\title{Lp Distance}
\usage{
lp_distance(x, y, p)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}

\item{p}{A positive integer indicating the order of the distance.}
}
\value{
The Lp distance between \code{x} and \code{y}.
}
\description{
Computes the Lp distance between two vectors.
}
\examples{
lp_distance(c(1, -2, 3), c(4, -5, 6), 3)
}
