% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_functionals.R
\name{lp_norm}
\alias{lp_norm}
\title{Lp Norm}
\usage{
lp_norm(x, p)
}
\arguments{
\item{x}{A numeric vector.}

\item{p}{A positive integer indicating the order of the norm.}
}
\value{
The Lp norm of the vector.
}
\description{
Computes the Lp norm of a vector.
}
\examples{
lp_norm(c(1, -2, 3), 3)
}
