% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_series.R
\name{max_bernoulli_b2n}
\alias{max_bernoulli_b2n}
\title{Maximum Bernoulli Number (B_{2n}) in Lookup Table}
\usage{
max_bernoulli_b2n()
}
\value{
The maximum index \(n\) for which Bernoulli numbers \(B_{2n}\) can be computed.
}
\description{
Returns the maximum index \(n\) for which Bernoulli numbers \(B_{2n}\) are precomputed.
}
