% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorials_and_binomial_coefficients.R
\name{max_factorial}
\alias{max_factorial}
\title{Maximum Factorial Lookup Value}
\usage{
max_factorial()
}
\value{
The maximum factorial value that can be computed.
}
\description{
Returns the maximum value for which factorials are precomputed.
}
