% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_series.R
\name{max_prime}
\alias{max_prime}
\title{Largest Prime Number in Lookup Table}
\usage{
max_prime()
}
\value{
The largest n-th prime number available in the lookup table.
}
\description{
Returns the largest n-th prime number that can be computed using the lookup table.
}
