% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_central_beta_distribution.R
\name{non_central_beta_cdf}
\alias{non_central_beta_cdf}
\title{Noncentral Beta Cumulative Distribution Function}
\usage{
non_central_beta_cdf(x, alpha, beta, lambda)
}
\arguments{
\item{x}{quantile}

\item{alpha}{first shape parameter (alpha > 0)}

\item{beta}{second shape parameter (beta > 0)}

\item{lambda}{noncentrality parameter (lambda >= 0)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Noncentral Beta Cumulative Distribution Function
}
\examples{
non_central_beta_cdf(0.5, 2, 3, 1)
}
