% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_central_beta_distribution.R
\name{non_central_beta_lpdf}
\alias{non_central_beta_lpdf}
\title{Noncentral Beta Log Probability Density Function}
\usage{
non_central_beta_lpdf(x, alpha, beta, lambda)
}
\arguments{
\item{x}{quantile}

\item{alpha}{first shape parameter (alpha > 0)}

\item{beta}{second shape parameter (beta > 0)}

\item{lambda}{noncentrality parameter (lambda >= 0)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Noncentral Beta Log Probability Density Function
}
\examples{
non_central_beta_lpdf(0.5, 2, 3, 1)
}
