% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_central_beta_distribution.R
\name{non_central_beta_quantile}
\alias{non_central_beta_quantile}
\title{Noncentral Beta Quantile Function}
\usage{
non_central_beta_quantile(p, alpha, beta, lambda)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{alpha}{first shape parameter (alpha > 0)}

\item{beta}{second shape parameter (beta > 0)}

\item{lambda}{noncentrality parameter (lambda >= 0)}
}
\value{
Quantile function value at \code{p}
}
\description{
Noncentral Beta Quantile Function
}
\examples{
non_central_beta_quantile(0.5, 2, 3, 1)
}
