% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_central_chi_squared_distribution.R
\name{non_central_chi_squared_cdf}
\alias{non_central_chi_squared_cdf}
\title{Noncentral Chi-Squared Cumulative Distribution Function}
\usage{
non_central_chi_squared_cdf(x, df, lambda)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (df > 0)}

\item{lambda}{noncentrality parameter (lambda >= 0)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Noncentral Chi-Squared Cumulative Distribution Function
}
\examples{
non_central_chi_squared_cdf(2, 3, 1)
}
