% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_central_chi_squared_distribution.R
\name{non_central_chi_squared_pdf}
\alias{non_central_chi_squared_pdf}
\title{Noncentral Chi-Squared Probability Density Function}
\usage{
non_central_chi_squared_pdf(x, df, lambda)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (df > 0)}

\item{lambda}{noncentrality parameter (lambda >= 0)}
}
\value{
Probability density function value at \code{x}
}
\description{
Noncentral Chi-Squared Probability Density Function
}
\examples{
non_central_chi_squared_pdf(2, 3, 1)
}
