% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_central_chi_squared_distribution.R
\name{non_central_chi_squared_quantile}
\alias{non_central_chi_squared_quantile}
\title{Noncentral Chi-Squared Quantile Function}
\usage{
non_central_chi_squared_quantile(p, df, lambda)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{df}{degrees of freedom (df > 0)}

\item{lambda}{noncentrality parameter (lambda >= 0)}
}
\value{
Quantile function value at \code{p}
}
\description{
Noncentral Chi-Squared Quantile Function
}
\examples{
non_central_chi_squared_quantile(0.5, 3, 1)
}
