% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_central_t_distribution.R
\name{non_central_t_cdf}
\alias{non_central_t_cdf}
\title{Noncentral T Cumulative Distribution Function}
\usage{
non_central_t_cdf(x, df, delta)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (df > 0)}

\item{delta}{noncentrality parameter (delta >= 0)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Noncentral T Cumulative Distribution Function
}
\examples{
non_central_t_cdf(0, 3, 1)
}
