% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_central_t_distribution.R
\name{non_central_t_distribution}
\alias{non_central_t_distribution}
\alias{non_central_t_pdf}
\alias{non_central_t_lpdf}
\alias{non_central_t_cdf}
\alias{non_central_t_lcdf}
\alias{non_central_t_quantile}
\title{Noncentral T Distribution Functions}
\usage{
non_central_t_distribution(df, delta)

non_central_t_pdf(x, df, delta)

non_central_t_lpdf(x, df, delta)

non_central_t_cdf(x, df, delta)

non_central_t_lcdf(x, df, delta)

non_central_t_quantile(p, df, delta)
}
\arguments{
\item{df}{degrees of freedom (df > 0)}

\item{delta}{noncentrality parameter (delta >= 0)}

\item{x}{quantile}

\item{p}{probability (0 <= p <= 1)}
}
\value{
A single numeric value with the computed probability density, log-probability density, cumulative distribution, log-cumulative distribution, or quantile depending on the function called.
}
\description{
Functions to compute the probability density function, cumulative distribution function, and quantile function for the Noncentral T distribution.

The noncentral T distribution is a generalization of the Student's t Distribution.

The PDF is:

\deqn{f(x;\nu;\delta) = \frac{\nu^{\nu/2}\,\nu!} {2^{\nu} e^{\delta^{2}/2} (\nu+x^{2})^{\nu/2}\Gamma\!\left(\frac{\nu}{2}\right)} \left( \frac{\sqrt{2}\,\delta x\, {}_1F_1\!\left(\frac{\nu}{2}+1;\frac{3}{2}; \frac{\delta^{2}x^{2}}{2(\nu+x^{2})}\right)} {(\nu+x^{2})\Gamma\!\left(\frac{\nu+1}{2}\right)} + \frac{ {}_1F_1\!\left(\frac{\nu+1}{2};\frac{1}{2}; \frac{\delta^{2}x^{2}}{2(\nu+x^{2})}\right)} {\sqrt{\nu+x^{2}}\Gamma\!\left(\frac{\nu}{2}+1\right)} \right)}

The CDF is:

\deqn{F(t;\nu;\delta) = \Phi(-\delta) + \frac{1}{2} \sum_{i=0}^{\infty} \left( P_i\, I_x\!\left(i+\frac{1}{2},\frac{\nu}{2}\right) + \frac{\delta}{\sqrt{2}}\, Q_i\, I_x\!\left(i+1,\frac{\nu}{2}\right) \right)}

Where:

\deqn{P_i = e^{-\delta^{2}/2} \frac{\left(\delta^{2}/2\right)^i}{i!}, \qquad Q_i = e^{-\delta^{2}/2} \frac{\left(\delta^{2}/2\right)^i}{\Gamma\!\left(i+\frac{3}{2}\right)}, \qquad x = \frac{t^{2}}{\nu+t^{2}}}
}
\examples{
# Noncentral T distribution with 5 degrees of freedom and noncentrality parameter 1
dist <- non_central_t_distribution(5, 1)
# Apply generic functions
cdf(dist, 0.5)
logcdf(dist, 0.5)
pdf(dist, 0.5)
logpdf(dist, 0.5)
hazard(dist, 0.5)
chf(dist, 0.5)
mean(dist)
median(dist)
mode(dist)
range(dist)
quantile(dist, 0.2)
standard_deviation(dist)
support(dist)
variance(dist)
skewness(dist)
kurtosis(dist)
kurtosis_excess(dist)

# Convenience functions
non_central_t_pdf(0, 5, 1)
non_central_t_lpdf(0, 5, 1)
non_central_t_cdf(0, 5, 1)
non_central_t_lcdf(0, 5, 1)
non_central_t_quantile(0.5, 5, 1)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/dist_ref/dists/nc_t_dist.html}{Boost Documentation} for more details on the mathematical background.
}
