% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_central_t_distribution.R
\name{non_central_t_lcdf}
\alias{non_central_t_lcdf}
\title{Noncentral T Log Cumulative Distribution Function}
\usage{
non_central_t_lcdf(x, df, delta)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (df > 0)}

\item{delta}{noncentrality parameter (delta >= 0)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Noncentral T Log Cumulative Distribution Function
}
\examples{
non_central_t_lcdf(0, 3, 1)
}
