% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_central_t_distribution.R
\name{non_central_t_lpdf}
\alias{non_central_t_lpdf}
\title{Noncentral T Log Probability Density Function}
\usage{
non_central_t_lpdf(x, df, delta)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (df > 0)}

\item{delta}{noncentrality parameter (delta >= 0)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Noncentral T Log Probability Density Function
}
\examples{
non_central_t_lpdf(0, 3, 1)
}
