% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_central_t_distribution.R
\name{non_central_t_pdf}
\alias{non_central_t_pdf}
\title{Noncentral T Probability Density Function}
\usage{
non_central_t_pdf(x, df, delta)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (df > 0)}

\item{delta}{noncentrality parameter (delta >= 0)}
}
\value{
Probability density function value at \code{x}
}
\description{
Noncentral T Probability Density Function
}
\examples{
non_central_t_pdf(0, 3, 1)
}
