% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_central_t_distribution.R
\name{non_central_t_quantile}
\alias{non_central_t_quantile}
\title{Noncentral T Quantile Function}
\usage{
non_central_t_quantile(p, df, delta)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{df}{degrees of freedom (df > 0)}

\item{delta}{noncentrality parameter (delta >= 0)}
}
\value{
Quantile function value at \code{p}
}
\description{
Noncentral T Quantile Function
}
\examples{
non_central_t_quantile(0.5, 3, 1)
}
