% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_distribution.R
\name{normal_cdf}
\alias{normal_cdf}
\title{Normal Cumulative Distribution Function}
\usage{
normal_cdf(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{quantile}

\item{mean}{mean parameter (default is 0)}

\item{sd}{standard deviation parameter (default is 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Normal Cumulative Distribution Function
}
\examples{
normal_cdf(0)
}
