% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_distribution.R
\name{normal_lcdf}
\alias{normal_lcdf}
\title{Normal Log Cumulative Distribution Function}
\usage{
normal_lcdf(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{quantile}

\item{mean}{mean parameter (default is 0)}

\item{sd}{standard deviation parameter (default is 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Normal Log Cumulative Distribution Function
}
\examples{
normal_lcdf(0)
}
