% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_distribution.R
\name{normal_lpdf}
\alias{normal_lpdf}
\title{Normal Log Probability Density Function}
\usage{
normal_lpdf(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{quantile}

\item{mean}{mean parameter (default is 0)}

\item{sd}{standard deviation parameter (default is 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Normal Log Probability Density Function
}
\examples{
normal_lpdf(0)
}
