% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_distribution.R
\name{normal_pdf}
\alias{normal_pdf}
\title{Normal Probability Density Function}
\usage{
normal_pdf(x, mean = 0, sd = 1)
}
\arguments{
\item{x}{quantile}

\item{mean}{mean parameter (default is 0)}

\item{sd}{standard deviation parameter (default is 1)}
}
\value{
Probability density function value at \code{x}
}
\description{
Normal Probability Density Function
}
\examples{
normal_pdf(0)
}
