% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_distribution.R
\name{normal_quantile}
\alias{normal_quantile}
\title{Normal Quantile Function}
\usage{
normal_quantile(p, mean = 0, sd = 1)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{mean}{mean parameter (default is 0)}

\item{sd}{standard deviation parameter (default is 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Normal Quantile Function
}
\examples{
normal_quantile(0.5)
}
