% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature_and_differentiation.R
\name{ooura_fourier_sin}
\alias{ooura_fourier_sin}
\title{Ooura Fourier Sine Integral}
\usage{
ooura_fourier_sin(
  f,
  omega = 1,
  relative_error_tolerance = sqrt(.Machine$double.eps),
  levels = 8
)
}
\arguments{
\item{f}{A function to integrate. It should accept a single numeric value and
return a single numeric value.}

\item{omega}{The frequency parameter for the sine integral.}

\item{relative_error_tolerance}{The relative error tolerance for the approximation.}

\item{levels}{The number of levels of refinement to apply. Default is 8.}
}
\value{
The approximate value of the Fourier sine integral.
}
\description{
This function computes the Fourier sine integral using Ooura's method.
}
