% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owens_t_function.R
\name{owens_t}
\alias{owens_t}
\title{Owens T Function}
\usage{
owens_t(h, a)
}
\arguments{
\item{h}{The first argument of the Owens T function (boundary parameter)}

\item{a}{The second argument of the Owens T function (slope parameter)}
}
\value{
The value of the Owens T function at (h, a).
}
\description{
Computes Owen's T function T(h, a), which gives the probability of the event
(X > h and 0 < Y < a*X) where X and Y are independent standard normal random variables.

\deqn{T(h,a)=\frac{1}{2\pi}\int_{0}^{a}\frac{\exp\!\left\{-\tfrac{1}{2}h^{2}(1+x^{2})\right\}}{1+x^{2}}\,dx,\quad (-\infty<h,a<+\infty)}
}
\examples{
# Owens T Function
owens_t(1, 0.5)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/owens_t.html}{Boost Documentation} for more details on the mathematical background.
}
