% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_distribution.R
\name{pareto_cdf}
\alias{pareto_cdf}
\title{Pareto Cumulative Distribution Function}
\usage{
pareto_cdf(x, shape = 1, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{shape}{shape parameter (default is 1)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Pareto Cumulative Distribution Function
}
\examples{
pareto_cdf(1)
}
