% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_distribution.R
\name{pareto_lpdf}
\alias{pareto_lpdf}
\title{Pareto Log Probability Density Function}
\usage{
pareto_lpdf(x, shape = 1, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{shape}{shape parameter (default is 1)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Pareto Log Probability Density Function
}
\examples{
pareto_lpdf(1)
}
