% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_distribution.R
\name{pareto_pdf}
\alias{pareto_pdf}
\title{Pareto Probability Density Function}
\usage{
pareto_pdf(x, shape = 1, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{shape}{shape parameter (default is 1)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Probability density function value at \code{x}
}
\description{
Pareto Probability Density Function
}
\examples{
pareto_pdf(1)
}
