% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_distribution.R
\name{pareto_quantile}
\alias{pareto_quantile}
\title{Pareto Quantile Function}
\usage{
pareto_quantile(p, shape = 1, scale = 1)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{shape}{shape parameter (default is 1)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Pareto Quantile Function
}
\examples{
pareto_quantile(0.5)
}
