% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_distribution.R
\name{poisson_cdf}
\alias{poisson_cdf}
\title{Poisson Cumulative Distribution Function}
\usage{
poisson_cdf(x, lambda = 1)
}
\arguments{
\item{x}{quantile}

\item{lambda}{rate parameter (default is 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Poisson Cumulative Distribution Function
}
\examples{
poisson_cdf(0, 1)
}
