% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_distribution.R
\name{poisson_distribution}
\alias{poisson_distribution}
\alias{poisson_pdf}
\alias{poisson_lpdf}
\alias{poisson_cdf}
\alias{poisson_lcdf}
\alias{poisson_quantile}
\title{Poisson Distribution Functions}
\usage{
poisson_distribution(lambda = 1)

poisson_pdf(x, lambda = 1)

poisson_lpdf(x, lambda = 1)

poisson_cdf(x, lambda = 1)

poisson_lcdf(x, lambda = 1)

poisson_quantile(p, lambda = 1)
}
\arguments{
\item{lambda}{rate parameter (default is 1)}

\item{x}{quantile}

\item{p}{probability (0 <= p <= 1)}
}
\value{
A single numeric value with the computed probability density, log-probability density, cumulative distribution, log-cumulative distribution, or quantile depending on the function called.
}
\description{
Functions to compute the probability density function, cumulative distribution function, and quantile function for the Poisson distribution.

The Poisson distribution expresses the probability of a number of events (or failures, arrivals, occurrences ...) occurring in a fixed period of time, provided these events occur with a known mean rate \eqn{\lambda} (events/time), and are independent of the time since the last event.

It has the Probability Mass Function:
\deqn{f(k;\lambda) = \frac{\lambda^k e^{-\lambda}}{k!}}
for \eqn{k} events, with an expected number of events \eqn{\lambda}.

\strong{Accuracy and Implementation Notes:}
The Poisson distribution is implemented in terms of the incomplete gamma functions (\code{gamma_p} and \code{gamma_q}) and as such should have low error rates.
The quantile function will by default return an integer result that has been rounded outwards to ensure that if an X\% quantile is requested, then at least the requested coverage will be present in the central region, and no more than the requested coverage will be present in the tails.
}
\examples{
# Poisson distribution with lambda = 1
dist <- poisson_distribution(1)
# Apply generic functions
cdf(dist, 5)
logcdf(dist, 5)
pdf(dist, 5)
logpdf(dist, 5)
hazard(dist, 5)
chf(dist, 5)
mean(dist)
median(dist)
mode(dist)
range(dist)
quantile(dist, 0.2)
standard_deviation(dist)
support(dist)
variance(dist)
skewness(dist)
kurtosis(dist)
kurtosis_excess(dist)

# Convenience functions
poisson_pdf(0, 1)
poisson_lpdf(0, 1)
poisson_cdf(0, 1)
poisson_lcdf(0, 1)
poisson_quantile(0.5, 1)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/dist_ref/dists/poisson_dist.html}{Boost Documentation} for more details on the mathematical background.
}
