% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_distribution.R
\name{poisson_lcdf}
\alias{poisson_lcdf}
\title{Poisson Log Cumulative Distribution Function}
\usage{
poisson_lcdf(x, lambda = 1)
}
\arguments{
\item{x}{quantile}

\item{lambda}{rate parameter (default is 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Poisson Log Cumulative Distribution Function
}
\examples{
poisson_lcdf(0, 1)
}
