% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_distribution.R
\name{poisson_lpdf}
\alias{poisson_lpdf}
\title{Poisson Log Probability Density Function}
\usage{
poisson_lpdf(x, lambda = 1)
}
\arguments{
\item{x}{quantile}

\item{lambda}{rate parameter (default is 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Poisson Log Probability Density Function
}
\examples{
poisson_lpdf(0, 1)
}
