% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_distribution.R
\name{poisson_pdf}
\alias{poisson_pdf}
\title{Poisson Probability Density Function}
\usage{
poisson_pdf(x, lambda = 1)
}
\arguments{
\item{x}{quantile}

\item{lambda}{rate parameter (default is 1)}
}
\value{
Probability density function value at \code{x}
}
\description{
Poisson Probability Density Function
}
\examples{
poisson_pdf(0, 1)
}
