% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_distribution.R
\name{poisson_quantile}
\alias{poisson_quantile}
\title{Poisson Quantile Function}
\usage{
poisson_quantile(p, lambda = 1)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{lambda}{rate parameter (default is 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Poisson Quantile Function
}
\examples{
poisson_quantile(0.5, 1)
}
