% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{polygamma}
\alias{polygamma}
\title{Polygamma Function}
\usage{
polygamma(n, z)
}
\arguments{
\item{n}{Order of the polygamma function (non-negative integer)}

\item{z}{Input numeric value}
}
\value{
The value of the polygamma function at z of order n.
}
\description{
Computes the polygamma function, which is the n-th logarithmic derivative of the gamma function.
}
