% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{powm1}
\alias{powm1}
\title{Power of x raised to the power of y minus 1}
\usage{
powm1(x, y)
}
\arguments{
\item{x}{Input numeric value}

\item{y}{Exponent value}
}
\value{
The value of x raised to the power of y minus 1.
}
\description{
Computes x raised to the power of y minus 1.
}
