% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_series.R
\name{prime}
\alias{prime}
\title{Lookup Table of Prime Numbers}
\usage{
prime(n)
}
\arguments{
\item{n}{Index of the prime number (must be a positive integer)}
}
\value{
The n-th prime number.
}
\description{
Computes the n-th prime number using a lookup table.
}
