% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootfinding_and_minimisation.R
\name{quadratic_roots}
\alias{quadratic_roots}
\title{Finding Quadratic Roots}
\usage{
quadratic_roots(a, b, c)
}
\arguments{
\item{a}{Coefficient of the quadratic term.}

\item{b}{Coefficient of the linear term.}

\item{c}{Constant term.}
}
\value{
A two-element numeric vector containing the two roots of the quadratic equation.
}
\description{
Finding Quadratic Roots
}
\examples{
quadratic_roots(1, -3, 2)
}
