% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootfinding_and_minimisation.R
\name{quartic_roots}
\alias{quartic_roots}
\title{Finding Quartic Roots}
\usage{
quartic_roots(a, b, c, d, e)
}
\arguments{
\item{a}{Coefficient of the quartic term.}

\item{b}{Coefficient of the cubic term.}

\item{c}{Coefficient of the quadratic term.}

\item{d}{Coefficient of the linear term.}

\item{e}{Constant term.}
}
\value{
A numeric vector containing the four roots of the quartic equation.
}
\description{
Finding Quartic Roots
}
\examples{
quartic_roots(1, -10, 35, -50, 24)
}
