% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayleigh_distribution.R
\name{rayleigh_cdf}
\alias{rayleigh_cdf}
\title{Rayleigh Cumulative Distribution Function}
\usage{
rayleigh_cdf(x, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{scale}{scale parameter (default is 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Rayleigh Cumulative Distribution Function
}
\examples{
rayleigh_cdf(1)
}
