% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayleigh_distribution.R
\name{rayleigh_lcdf}
\alias{rayleigh_lcdf}
\title{Rayleigh Log Cumulative Distribution Function}
\usage{
rayleigh_lcdf(x, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Rayleigh Log Cumulative Distribution Function
}
\examples{
rayleigh_lcdf(1)
}
