% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayleigh_distribution.R
\name{rayleigh_lpdf}
\alias{rayleigh_lpdf}
\title{Rayleigh Log Probability Density Function}
\usage{
rayleigh_lpdf(x, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Rayleigh Log Probability Density Function
}
\examples{
rayleigh_lpdf(1)
}
