% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayleigh_distribution.R
\name{rayleigh_pdf}
\alias{rayleigh_pdf}
\title{Rayleigh Probability Density Function}
\usage{
rayleigh_pdf(x, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{scale}{scale parameter (default is 1)}
}
\value{
Probability density function value at \code{x}
}
\description{
Rayleigh Probability Density Function
}
\examples{
rayleigh_pdf(1)
}
