% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayleigh_distribution.R
\name{rayleigh_quantile}
\alias{rayleigh_quantile}
\title{Rayleigh Quantile Function}
\usage{
rayleigh_quantile(p, scale = 1)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Rayleigh Quantile Function
}
\examples{
rayleigh_quantile(0.5)
}
