% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorials_and_binomial_coefficients.R
\name{rising_factorial}
\alias{rising_factorial}
\title{Rising Factorial}
\usage{
rising_factorial(x, i)
}
\arguments{
\item{x}{Base value}

\item{i}{Exponent (non-negative integer)}
}
\value{
The value of the rising factorial.
}
\description{
Computes the rising factorial (also known as Pochhammer symbol) of a number.
}
