% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{rsqrt}
\alias{rsqrt}
\title{Inverse square root}
\usage{
rsqrt(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The inverse square root of x.
}
\description{
Computes the inverse square root of x.
}
