% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{sin_pi}
\alias{sin_pi}
\title{Sine of Pi}
\usage{
sin_pi(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The sine of x multiplied by pi.
}
\description{
Computes the sine of a value multiplied by pi.
}
