% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinus_cardinal_hyperbolic_functions.R
\name{sinc_pi}
\alias{sinc_pi}
\title{Sinus Cardinal Function}
\usage{
sinc_pi(x)
}
\arguments{
\item{x}{Input value}
}
\value{
The value of the sinus cardinal function \eqn{(\text{sinc}(x))}.
}
\description{
Computes the sinus cardinal function \eqn{(\text{sinc}(x) = \frac{\sin(x)}{x})}.
}
