% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature.R
\name{sinh_sinh}
\alias{sinh_sinh}
\title{Sinh-Sinh Quadrature for Numerical Integration}
\usage{
sinh_sinh(f, tol = .Machine$double.eps, max_refinements = 9)
}
\arguments{
\item{f}{A function to integrate. It should accept a single numeric value and
return a single numeric value.}

\item{tol}{The tolerance for the approximation. Default is \code{.Machine$double.eps}.}

\item{max_refinements}{The maximum number of refinements to apply. Default is 9.}
}
\value{
The approximate value of the integral.
}
\description{
This function performs integration using a sinh-sinh quadrature routine.
}
