% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinus_cardinal_hyperbolic_functions.R
\name{sinhc_pi}
\alias{sinhc_pi}
\title{Hyperbolic Sinus Cardinal Function}
\usage{
sinhc_pi(x)
}
\arguments{
\item{x}{Input value}
}
\value{
The value of the hyperbolic sinus cardinal function \eqn{(\text{sinhc}(x))}.
}
\description{
Computes the hyperbolic sinus cardinal function \eqn{(\text{sinhc}(x) = \frac{\sinh(x)}{x})}.
}
