% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_normal_distribution.R
\name{skew_normal_cdf}
\alias{skew_normal_cdf}
\title{Skew Normal Cumulative Distribution Function}
\usage{
skew_normal_cdf(x, location = 0, scale = 1, shape = 0)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}

\item{shape}{shape parameter (default is 0)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Skew Normal Cumulative Distribution Function
}
\examples{
skew_normal_cdf(0)
}
