% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_normal_distribution.R
\name{skew_normal_distribution}
\alias{skew_normal_distribution}
\alias{skew_normal_pdf}
\alias{skew_normal_lpdf}
\alias{skew_normal_cdf}
\alias{skew_normal_lcdf}
\alias{skew_normal_quantile}
\title{Skew Normal Distribution Functions}
\usage{
skew_normal_distribution(location = 0, scale = 1, shape = 0)

skew_normal_pdf(x, location = 0, scale = 1, shape = 0)

skew_normal_lpdf(x, location = 0, scale = 1, shape = 0)

skew_normal_cdf(x, location = 0, scale = 1, shape = 0)

skew_normal_lcdf(x, location = 0, scale = 1, shape = 0)

skew_normal_quantile(p, location = 0, scale = 1, shape = 0)
}
\arguments{
\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}

\item{shape}{shape parameter (default is 0)}

\item{x}{quantile}

\item{p}{probability (0 <= p <= 1)}
}
\value{
A single numeric value with the computed probability density, log-probability density, cumulative distribution, log-cumulative distribution, or quantile depending on the function called.
}
\description{
Functions to compute the probability density function, cumulative distribution function, and quantile function for the Skew Normal distribution.

The skew normal distribution is a variant of the most well known Gaussian statistical distribution.
If the standard (mean = 0, scale = 1) normal distribution probability density function is \eqn{\phi(x)} and the cumulative distribution function is \eqn{\Phi(x)},
then the PDF of the skew normal distribution with shape parameter \eqn{\alpha} is:
\deqn{f(x;\alpha) = 2\phi(x)\Phi(\alpha x)}

Given location \eqn{\xi}, scale \eqn{\omega}, and shape \eqn{\alpha}, it can be transformed to:
\deqn{f(x) = \frac{2}{\omega}\phi\left(\frac{x-\xi}{\omega}\right)\Phi\left(\alpha\frac{x-\xi}{\omega}\right)}

\strong{Accuracy and Implementation Notes:}
The skew_normal distribution with shape = zero is equivalent to the normal distribution and uses the error function for excellent accuracy.
The CDF requires Owen's T function, which is evaluated using algoritms by Patefield and Tandy.
The median and mode are calculated by iterative root finding and may be less accurate than other estimates.
}
\examples{
# Skew Normal distribution with location = 0, scale = 1, shape = 0
dist <- skew_normal_distribution(0, 1, 0)
# Apply generic functions
cdf(dist, 0.5)
logcdf(dist, 0.5)
pdf(dist, 0.5)
logpdf(dist, 0.5)
hazard(dist, 0.5)
chf(dist, 0.5)
mean(dist)
median(dist)
mode(dist)
range(dist)
quantile(dist, 0.2)
standard_deviation(dist)
support(dist)
variance(dist)
skewness(dist)
kurtosis(dist)
kurtosis_excess(dist)

# Convenience functions
skew_normal_pdf(0)
skew_normal_lpdf(0)
skew_normal_cdf(0)
skew_normal_lcdf(0)
skew_normal_quantile(0.5)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/dist_ref/dists/skew_normal_dist.html}{Boost Documentation} for more details on the mathematical background.
}
