% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_normal_distribution.R
\name{skew_normal_pdf}
\alias{skew_normal_pdf}
\title{Skew Normal Probability Density Function}
\usage{
skew_normal_pdf(x, location = 0, scale = 1, shape = 0)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}

\item{shape}{shape parameter (default is 0)}
}
\value{
Probability density function value at \code{x}
}
\description{
Skew Normal Probability Density Function
}
\examples{
skew_normal_pdf(0)
}
