% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew_normal_distribution.R
\name{skew_normal_quantile}
\alias{skew_normal_quantile}
\title{Skew Normal Quantile Function}
\usage{
skew_normal_quantile(p, location = 0, scale = 1, shape = 0)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}

\item{shape}{shape parameter (default is 0)}
}
\value{
Quantile function value at \code{p}
}
\description{
Skew Normal Quantile Function
}
\examples{
skew_normal_quantile(0.5)
}
