% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_functions.R
\name{sph_bessel}
\alias{sph_bessel}
\title{Spherical Bessel Function of the First Kind}
\usage{
sph_bessel(v, x)
}
\arguments{
\item{v}{Order of the spherical Bessel function}

\item{x}{Argument of the spherical Bessel function}
}
\value{
The value of the spherical Bessel function of the first kind j_v(x).
}
\description{
Calculates the spherical Bessel function of the first kind j_v(x).
}
