% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_functions.R
\name{sph_bessel_prime}
\alias{sph_bessel_prime}
\title{Derivative of the Spherical Bessel Function of the First Kind}
\usage{
sph_bessel_prime(v, x)
}
\arguments{
\item{v}{Order of the spherical Bessel function}

\item{x}{Argument of the spherical Bessel function}
}
\value{
The value of the derivative of the spherical Bessel function of the first kind j_v(x).
}
\description{
Calculates the derivative of the spherical Bessel function of the first kind j_v(x).
}
