% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bessel_functions.R
\name{sph_neumann}
\alias{sph_neumann}
\title{Spherical Bessel Function of the Second Kind}
\usage{
sph_neumann(v, x)
}
\arguments{
\item{v}{Order of the spherical Bessel function}

\item{x}{Argument of the spherical Bessel function}
}
\value{
The value of the spherical Bessel function of the second kind y_v(x).
}
\description{
Calculates the spherical Bessel function of the second kind y_v(x).
}
