% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{spherical_harmonic_i}
\alias{spherical_harmonic_i}
\title{Spherical Harmonic Function (Imaginary Part)}
\usage{
spherical_harmonic_i(n, m, theta, phi)
}
\arguments{
\item{n}{Degree of the spherical harmonic}

\item{m}{Order of the spherical harmonic}

\item{theta}{Polar angle (colatitude)}

\item{phi}{Azimuthal angle (longitude)}
}
\value{
The imaginary part of the spherical harmonic function \eqn{\(Y_{n}^{m}(\theta, \phi)\)}.
}
\description{
Computes the imaginary part of the spherical harmonic function \eqn{\(Y_{n}^{m}(\theta, \phi)\)}.
}
