% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{spherical_harmonic_r}
\alias{spherical_harmonic_r}
\title{Spherical Harmonic Function (Real Part)}
\usage{
spherical_harmonic_r(n, m, theta, phi)
}
\arguments{
\item{n}{Degree of the spherical harmonic}

\item{m}{Order of the spherical harmonic}

\item{theta}{Polar angle (colatitude)}

\item{phi}{Azimuthal angle (longitude)}
}
\value{
The real part of the spherical harmonic function \eqn{(Y_{n}^{m}(\theta, \phi))}.
}
\description{
Computes the real part of the spherical harmonic function \eqn{(Y_{n}^{m}(\theta, \phi))}.
}
