% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{sqrt1pm1}
\alias{sqrt1pm1}
\title{Square root of 1 plus x minus 1}
\usage{
sqrt1pm1(x)
}
\arguments{
\item{x}{Input numeric value}
}
\value{
The square root of 1 plus x minus 1.
}
\description{
Subtract 1 from the square root of 1 plus x.
}
