% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_t_distribution.R
\name{students_t_cdf}
\alias{students_t_cdf}
\title{Student's t Cumulative Distribution Function}
\usage{
students_t_cdf(x, df = 1)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (default is 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Student's t Cumulative Distribution Function
}
\examples{
students_t_cdf(0, 3)
}
