% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_t_distribution.R
\name{students_t_lcdf}
\alias{students_t_lcdf}
\title{Student's t Log Cumulative Distribution Function}
\usage{
students_t_lcdf(x, df = 1)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (default is 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Student's t Log Cumulative Distribution Function
}
\examples{
students_t_lcdf(0, 3)
}
