% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_t_distribution.R
\name{students_t_lpdf}
\alias{students_t_lpdf}
\title{Student's t Log Probability Density Function}
\usage{
students_t_lpdf(x, df = 1)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (default is 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Student's t Log Probability Density Function
}
\examples{
students_t_lpdf(0, 3)
}
