% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_t_distribution.R
\name{students_t_pdf}
\alias{students_t_pdf}
\title{Student's t Probability Density Function}
\usage{
students_t_pdf(x, df = 1)
}
\arguments{
\item{x}{quantile}

\item{df}{degrees of freedom (default is 1)}
}
\value{
Probability density function value at \code{x}
}
\description{
Student's t Probability Density Function
}
\examples{
students_t_pdf(0, 3)
}
