% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students_t_distribution.R
\name{students_t_quantile}
\alias{students_t_quantile}
\title{Student's t Quantile Function}
\usage{
students_t_quantile(p, df = 1)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{df}{degrees of freedom (default is 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Student's t Quantile Function
}
\examples{
students_t_quantile(0.5, 3)
}
