% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_functionals.R
\name{sup_distance}
\alias{sup_distance}
\title{Supremum Distance}
\usage{
sup_distance(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector of the same length as \code{x}.}
}
\value{
The supremum distance between \code{x} and \code{y}.
}
\description{
Computes the supremum distance (maximum distance) between two vectors.
}
\examples{
sup_distance(c(1, -2, 3), c(4, -5, 6))
}
