% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_functionals.R
\name{sup_norm}
\alias{sup_norm}
\title{Supremum Norm}
\usage{
sup_norm(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The supremum norm of the vector.
}
\description{
Computes the supremum norm (maximum norm) of a vector.
}
\examples{
sup_norm(c(1, -2, 3))
}
