% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_series.R
\name{tangent_t2n_range}
\alias{tangent_t2n_range}
\title{Tangent Numbers (\eqn{T_{2n}}) Range}
\usage{
tangent_t2n_range(start_index, number_of_tangent_t2n)
}
\arguments{
\item{start_index}{The starting index for the tangent numbers (must be a non-negative integer)}

\item{number_of_tangent_t2n}{The number of tangent numbers to compute}
}
\value{
A numeric vector containing the tangent numbers \eqn{(T_{2n})} for the specified range.
}
\description{
Computes a range of tangent numbers \eqn{(T_{2n})} starting from a given index.
}
