% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature.R
\name{tanh_sinh}
\alias{tanh_sinh}
\title{Tanh-Sinh Quadrature for Numerical Integration}
\usage{
tanh_sinh(f, a, b, tol = .Machine$double.eps, max_refinements = 15)
}
\arguments{
\item{f}{A function to integrate. It should accept a single numeric value and}

\item{a}{The lower limit of integration.}

\item{b}{The upper limit of integration.}

\item{tol}{The tolerance for the approximation. Default is \code{.Machine$double.eps}.}

\item{max_refinements}{The maximum number of refinements to apply. Default is 15.}
}
\value{
The approximate value of the integral.
}
\description{
This function performs integration using a tanh-sinh quadrature routine.
}
