% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{tgamma}
\alias{tgamma}
\title{Gamma Function}
\usage{
tgamma(z)
}
\arguments{
\item{z}{Input numeric value}
}
\value{
The value of the gamma function at z.
}
\description{
Computes the gamma function for a given input.
}
