% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{tgamma1pm1}
\alias{tgamma1pm1}
\title{Gamma Function (1 + z) - 1}
\usage{
tgamma1pm1(dz)
}
\arguments{
\item{dz}{Input numeric value (1 + z)}
}
\value{
The value of the gamma function at (1 + dz) - 1
}
\description{
Computes the gamma function for (1 + z) - 1, which is useful for small values of z.
}
