% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{tgamma_delta_ratio}
\alias{tgamma_delta_ratio}
\title{Ratio of Gamma Functions (Delta input)}
\usage{
tgamma_delta_ratio(a, delta)
}
\arguments{
\item{a}{Numerator argument for the gamma function}

\item{delta}{Increment added to the numerator argument}
}
\value{
The ratio of the gamma functions gamma(a) / gamma(a + delta).
}
\description{
Computes the ratio of two gamma functions, specifically gamma(a) / gamma(a + delta).
}
