% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{tgamma_lower}
\alias{tgamma_lower}
\title{Full (Non-normalised) Lower Incomplete Gamma Function}
\usage{
tgamma_lower(a, z)
}
\arguments{
\item{a}{Shape parameter of the gamma function}

\item{z}{Upper limit of integration}
}
\value{
The value of the full lower incomplete gamma function gamma(a, z).
}
\description{
Computes the full lower incomplete gamma function gamma(a, z).
}
