% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{tgamma_ratio}
\alias{tgamma_ratio}
\title{Ratio of Gamma Functions}
\usage{
tgamma_ratio(a, b)
}
\arguments{
\item{a}{Numerator argument for the gamma function}

\item{b}{Denominator argument for the gamma function}
}
\value{
The ratio of the gamma functions gamma(a) / gamma(b).
}
\description{
Computes the ratio of two gamma functions, specifically gamma(a) / gamma(b).
}
