% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{tgamma_upper}
\alias{tgamma_upper}
\title{Full (Non-normalised) Upper Incomplete Gamma Function}
\usage{
tgamma_upper(a, z)
}
\arguments{
\item{a}{Shape parameter of the gamma function}

\item{z}{Upper limit of integration}
}
\value{
The value of the full upper incomplete gamma function gamma(a, z).
}
\description{
Computes the full upper incomplete gamma function gamma(a, z).
}
