% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rootfinding_and_minimisation.R
\name{toms748_solve}
\alias{toms748_solve}
\title{Root-Finding with the TOMS 748 Algorithm}
\usage{
toms748_solve(
  f,
  lower,
  upper,
  digits = .Machine$double.digits,
  max_iter = .Machine$integer.max
)
}
\arguments{
\item{f}{A function to find the root of. It should take and return a single numeric value.}

\item{lower}{The lower bound of the interval to search for the root.}

\item{upper}{The upper bound of the interval to search for the root.}

\item{digits}{The number of significant digits to which the root should be found. Defaults to double precision.}

\item{max_iter}{The maximum number of iterations to perform. Defaults to the maximum integer value.}
}
\value{
A three-element list containing:
\itemize{
\item \code{lower}: The lower bound of the interval containing the root.
\item \code{upper}: The upper bound of the interval containing the root.
\item \code{iterations}: The number of iterations performed.
}
}
\description{
Root-Finding with the TOMS 748 Algorithm
}
\examples{
f <- function(x) x^2 - 2
toms748_solve(f, lower = 0, upper = 2)
}
