% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_functionals.R
\name{total_variation}
\alias{total_variation}
\title{Total Variation}
\usage{
total_variation(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The total variation of the vector.
}
\description{
Computes the total variation of a numeric vector.
}
\examples{
total_variation(c(1, 2, 1, 3))
}
