% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature_and_differentiation.R
\name{trapezoidal}
\alias{trapezoidal}
\title{Trapezoidal Rule for Numerical Integration}
\usage{
trapezoidal(f, a, b, tol = .Machine$double.eps, max_refinements = 12)
}
\arguments{
\item{f}{A function to integrate. It should accept a single numeric value and
return a single numeric value.}

\item{a}{The lower limit of integration.}

\item{b}{The upper limit of integration.}

\item{tol}{The tolerance for the approximation. Default is \code{.Machine$double.eps}.}

\item{max_refinements}{The maximum number of refinements to apply. Default is 12.}
}
\value{
The approximate value of the integral.
}
\description{
This function approximates the integral of a function using the trapezoidal rule.
}
