% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular_distribution.R
\name{triangular_lcdf}
\alias{triangular_lcdf}
\title{Triangular Log Cumulative Distribution Function}
\usage{
triangular_lcdf(x, lower = 0, mode = 1, upper = 2)
}
\arguments{
\item{x}{quantile}

\item{lower}{lower limit of the distribution (default is 0)}

\item{mode}{mode of the distribution (default is 1)}

\item{upper}{upper limit of the distribution (default is 2)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Triangular Log Cumulative Distribution Function
}
\examples{
triangular_lcdf(1)
}
