% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular_distribution.R
\name{triangular_pdf}
\alias{triangular_pdf}
\title{Triangular Probability Density Function}
\usage{
triangular_pdf(x, lower = 0, mode = 1, upper = 2)
}
\arguments{
\item{x}{quantile}

\item{lower}{lower limit of the distribution (default is 0)}

\item{mode}{mode of the distribution (default is 1)}

\item{upper}{upper limit of the distribution (default is 2)}
}
\value{
Probability density function value at \code{x}
}
\description{
Triangular Probability Density Function
}
\examples{
triangular_pdf(1)
}
