% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular_distribution.R
\name{triangular_quantile}
\alias{triangular_quantile}
\title{Triangular Quantile Function}
\usage{
triangular_quantile(p, lower = 0, mode = 1, upper = 2)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{lower}{lower limit of the distribution (default is 0)}

\item{mode}{mode of the distribution (default is 1)}

\item{upper}{upper limit of the distribution (default is 2)}
}
\value{
Quantile function value at \code{p}
}
\description{
Triangular Quantile Function
}
\examples{
triangular_quantile(0.5)
}
