% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_functions.R
\name{trigamma_boost}
\alias{trigamma_boost}
\title{Trigamma Function}
\usage{
trigamma_boost(z)
}
\arguments{
\item{z}{Input numeric value}
}
\value{
The value of the trigamma function at z.
}
\description{
Computes the trigamma function (the second logarithmic derivative of the gamma function) for a given input.
}
