% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorials_and_binomial_coefficients.R
\name{unchecked_factorial}
\alias{unchecked_factorial}
\title{Factorial without checks (table lookup)}
\usage{
unchecked_factorial(i)
}
\arguments{
\item{i}{Non-negative integer input}
}
\value{
The factorial of the input integer.
}
\description{
Computes the factorial of a non-negative integer without input validation
using a table lookup for performance.
}
