% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_distribution.R
\name{uniform_cdf}
\alias{uniform_cdf}
\title{Uniform Cumulative Distribution Function}
\usage{
uniform_cdf(x, lower = 0, upper = 1)
}
\arguments{
\item{x}{quantile}

\item{lower}{lower bound of the distribution (default is 0)}

\item{upper}{upper bound of the distribution (default is 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Uniform Cumulative Distribution Function
}
\examples{
uniform_cdf(0.5)
}
