% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_distribution.R
\name{uniform_lcdf}
\alias{uniform_lcdf}
\title{Uniform Log Cumulative Distribution Function}
\usage{
uniform_lcdf(x, lower = 0, upper = 1)
}
\arguments{
\item{x}{quantile}

\item{lower}{lower bound of the distribution (default is 0)}

\item{upper}{upper bound of the distribution (default is 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Uniform Log Cumulative Distribution Function
}
\examples{
uniform_lcdf(0.5)
}
