% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_distribution.R
\name{uniform_lpdf}
\alias{uniform_lpdf}
\title{Uniform Log Probability Density Function}
\usage{
uniform_lpdf(x, lower = 0, upper = 1)
}
\arguments{
\item{x}{quantile}

\item{lower}{lower bound of the distribution (default is 0)}

\item{upper}{upper bound of the distribution (default is 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Uniform Log Probability Density Function
}
\examples{
uniform_lpdf(0.5)
}
