% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_distribution.R
\name{uniform_pdf}
\alias{uniform_pdf}
\title{Uniform Probability Density Function}
\usage{
uniform_pdf(x, lower = 0, upper = 1)
}
\arguments{
\item{x}{quantile}

\item{lower}{lower bound of the distribution (default is 0)}

\item{upper}{upper bound of the distribution (default is 1)}
}
\value{
Probability density function value at \code{x}
}
\description{
Uniform Probability Density Function
}
\examples{
uniform_pdf(0.5)
}
