% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform_distribution.R
\name{uniform_quantile}
\alias{uniform_quantile}
\title{Uniform Quantile Function}
\usage{
uniform_quantile(p, lower = 0, upper = 1)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{lower}{lower bound of the distribution (default is 0)}

\item{upper}{upper bound of the distribution (default is 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Uniform Quantile Function
}
\examples{
uniform_quantile(0.5)
}
