% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_distribution.R
\name{weibull_cdf}
\alias{weibull_cdf}
\title{Weibull Cumulative Distribution Function}
\usage{
weibull_cdf(x, shape = 1, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{shape}{shape parameter (default is 1)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Cumulative distribution function value at \code{x}
}
\description{
Weibull Cumulative Distribution Function
}
\examples{
weibull_cdf(1)
}
