% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_distribution.R
\name{weibull_lcdf}
\alias{weibull_lcdf}
\title{Weibull Log Cumulative Distribution Function}
\usage{
weibull_lcdf(x, shape = 1, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{shape}{shape parameter (default is 1)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log cumulative distribution function value at \code{x}
}
\description{
Weibull Log Cumulative Distribution Function
}
\examples{
weibull_lcdf(1)
}
