% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_distribution.R
\name{weibull_lpdf}
\alias{weibull_lpdf}
\title{Weibull Log Probability Density Function}
\usage{
weibull_lpdf(x, shape = 1, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{shape}{shape parameter (default is 1)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Log probability density function value at \code{x}
}
\description{
Weibull Log Probability Density Function
}
\examples{
weibull_lpdf(1)
}
