% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_distribution.R
\name{weibull_pdf}
\alias{weibull_pdf}
\title{Weibull Probability Density Function}
\usage{
weibull_pdf(x, shape = 1, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{shape}{shape parameter (default is 1)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Probability density function value at \code{x}
}
\description{
Weibull Probability Density Function
}
\examples{
weibull_pdf(1)
}
