% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weibull_distribution.R
\name{weibull_quantile}
\alias{weibull_quantile}
\title{Weibull Quantile Function}
\usage{
weibull_quantile(p, shape = 1, scale = 1)
}
\arguments{
\item{p}{probability (0 <= p <= 1)}

\item{shape}{shape parameter (default is 1)}

\item{scale}{scale parameter (default is 1)}
}
\value{
Quantile function value at \code{p}
}
\description{
Weibull Quantile Function
}
\examples{
weibull_quantile(0.5)
}
