% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anderson_darling_test.R
\name{anderson_darling_test}
\alias{anderson_darling_test}
\alias{anderson_darling_normality_statistic}
\title{Anderson-Darling Test for Normality}
\usage{
anderson_darling_normality_statistic(x, mu = 0, sd = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{mu}{A single numeric value.}

\item{sd}{A single numeric value.}
}
\value{
A numeric value or vector with the computed statistic.
}
\description{
Functions to perform the Anderson-Darling test for normality.
}
\examples{
# Anderson-Darling test for normality
anderson_darling_normality_statistic(c(1, 2, 3, 4, 5), 0, 1)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/anderson_darling.html}{Boost Documentation} for more details on the mathematical background.
}
