% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostmath-package.R
\docType{package}
\name{boostmath-package}
\alias{boostmath}
\alias{boostmath-package}
\title{boostmath: 'R' Bindings for the 'Boost' Math Functions}
\description{
'R' bindings for the various functions and statistical distributions provided by the 'Boost' Math library \url{https://www.boost.org/doc/libs/boost_1_88_0/libs/math/doc/html/index.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrjohns/boostmath}
  \item \url{https://www.boost.org/doc/libs/boost_1_88_0/libs/math/doc/html/index.html}
  \item Report bugs at \url{https://github.com/andrjohns/boostmath/issues}
}

}
\author{
\strong{Maintainer}: Andrew R. Johnson \email{andrew.johnson@arjohnsonau.com} (\href{https://orcid.org/0000-0001-7000-8065}{ORCID})

}
\keyword{internal}
