% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chatterjee_correlation.R
\name{chatterjee_correlation}
\alias{chatterjee_correlation}
\title{Chatterjee Correlation Function}
\usage{
chatterjee_correlation(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A two-element numeric vector containing the test statistic and the p-value.
}
\description{
Functions to compute the Chatterjee correlation.
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(2, 3, 5, 7, 11)
# Chatterjee correlation
chatterjee_correlation(x, y)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/chatterjee_correlation.html}{Boost Documentation} for more details on the mathematical background.
}
