% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{chebyshev_t}
\alias{chebyshev_t}
\title{Chebyshev Polynomial of the First Kind}
\usage{
chebyshev_t(n, x)
}
\arguments{
\item{n}{Degree of the polynomial}

\item{x}{Argument of the polynomial}
}
\value{
The value of the Chebyshev polynomial \eqn{(T_n(x))}.
}
\description{
Computes the Chebyshev polynomial of the first kind \eqn{(T_n(x))}.
}
