% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\title{Boost Math Constants}
\usage{
constants(constant = NULL)
}
\arguments{
\item{constant}{A string specifying the name of the constant to retrieve. If \code{NULL}, returns a list of all constants (see documentation below for full list).}
}
\value{
Requested constant value if \code{constant} is specified, otherwise a list of all available constants.
}
\description{
Provides access to mathematical constants used in the Boost Math library.
}
\examples{
constants()
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/constants.html}{Boost Documentation} for more details on the constants.
}
