% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_integrals.R
\name{exponential_integrals}
\alias{exponential_integrals}
\alias{expint_en}
\alias{expint_ei}
\title{Exponential Integrals}
\usage{
expint_en(n, z)

expint_ei(z)
}
\arguments{
\item{n}{Order of the integral (for En)}

\item{z}{Argument of the integral (for En and Ei)}
}
\value{
A single numeric value with the computed exponential integral.
}
\description{
Functions to compute various exponential integrals, including En and Ei.

\strong{Exponential Integral En}
\itemize{
\item \code{expint_en(n, z)}: Exponential integral En of z: \eqn{E_n(z)}:
}

\deqn{E_n(z) = \int_{1}^\infty \frac{e^{-xt}}{t^n}dt}

\strong{Exponential Integral Ei}
\itemize{
\item \code{expint_ei(n, z)}: Exponential integral Ei of z: \eqn{E_i(z)}:
}

\deqn{E_i(z) = \int_{x}^\infty \frac{e^{-t}}{t}dt}s
}
\examples{
# Exponential integral En with n = 1 and z = 2
expint_en(1, 2)
# Exponential integral Ei with z = 2
expint_ei(2)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/expint.html}{Boost Documentation} for
}
