% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_integrals.R
\name{exponential_integrals}
\alias{exponential_integrals}
\alias{expint_en}
\alias{expint_ei}
\title{Exponential Integrals}
\usage{
expint_en(n, z)

expint_ei(z)
}
\arguments{
\item{n}{Order of the integral (for En), must be a non-negative integer}

\item{z}{Argument of the integral (for En and Ei)}
}
\value{
A single numeric value with the computed exponential integral.
}
\description{
Functions to compute various exponential integrals, including En and Ei.

\strong{Exponential Integral En:}

Defined by the integral:
\deqn{E_n(z) = \int_1^\infty \frac{e^{-zt}}{t^n} dt}

\strong{Exponential Integral Ei:}

Defined by the integral:
\deqn{Ei(z) = -\int_{-z}^\infty \frac{e^{-t}}{t} dt}
}
\examples{
# Exponential integral En with n = 1 and z = 2
expint_en(1, 2)
# Exponential integral Ei with z = 2
expint_ei(2)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/expint.html}{Boost Documentation}
for more details on the mathematical background.
}
