% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_integrals.R
\name{exponential_integrals}
\alias{exponential_integrals}
\alias{expint_en}
\alias{expint_ei}
\title{Exponential Integrals}
\usage{
expint_en(n, z)

expint_ei(z)
}
\arguments{
\item{n}{Order of the integral (for En)}

\item{z}{Argument of the integral (for En and Ei)}
}
\description{
Functions to compute various exponential integrals, including En and Ei.
}
\examples{
# Exponential integral En with n = 1 and z = 2
expint_en(1, 2)
# Exponential integral Ei with z = 2
expint_ei(2)
}
\seealso{
\href{https://www.boost.org/doc/libs/1_87_0/libs/math/doc/html/math_toolkit/special_functions/exponential_integrals.html}{Boost Documentation} for
}
