% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{hypot}
\alias{hypot}
\title{Hypotenuse}
\usage{
hypot(x)
}
\arguments{
\item{x}{Input numeric value representing one side of the triangle}
}
\value{
The length of the hypotenuse.
}
\description{
Computes the hypotenuse of a right triangle given the lengths of the two sides.
}
