% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverse_gaussian_distribution.R
\name{inverse_gaussian_distribution}
\alias{inverse_gaussian_distribution}
\alias{inverse_gaussian_pdf}
\alias{inverse_gaussian_lpdf}
\alias{inverse_gaussian_cdf}
\alias{inverse_gaussian_lcdf}
\alias{inverse_gaussian_quantile}
\title{Inverse Gaussian Distribution Functions}
\usage{
inverse_gaussian_pdf(x, mu, lambda)

inverse_gaussian_lpdf(x, mu, lambda)

inverse_gaussian_cdf(x, mu, lambda)

inverse_gaussian_lcdf(x, mu, lambda)

inverse_gaussian_quantile(p, mu, lambda)
}
\arguments{
\item{x}{quantile}

\item{mu}{mean parameter (mu > 0)}

\item{lambda}{scale (precision) parameter (lambda > 0)}

\item{p}{probability (0 <= p <= 1)}
}
\description{
Functions to compute the probability density function, cumulative distribution function, and quantile function for the Inverse Gaussian distribution.
}
\examples{
# Inverse Gaussian distribution with mu = 3, lambda = 4
inverse_gaussian_pdf(2, 3, 4)
inverse_gaussian_lpdf(2, 3, 4)
inverse_gaussian_cdf(2, 3, 4)
inverse_gaussian_lcdf(2, 3, 4)
inverse_gaussian_quantile(0.5, 3, 4)
}
\seealso{
\href{https://www.boost.org/doc/libs/1_87_0/libs/math/doc/html/math_toolkit/dist_ref/dists/inverse_gaussian_dist.html}{Boost Documentation} for more details on the mathematical background.
}
