% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ljung_box_test.R
\name{ljung_box_test}
\alias{ljung_box_test}
\alias{ljung_box}
\title{Ljung-Box Test for Autocorrelation}
\usage{
ljung_box(v, lags = -1, fit_dof = 0)
}
\arguments{
\item{v}{A numeric vector.}

\item{lags}{A single integer value (default uses \deqn{\log(n)}).}

\item{fit_dof}{A single integer value.}
}
\value{
A two-element numeric vector containing the test statistic and the p-value.
}
\description{
Functions to perform the Ljung-Box test for autocorrelation in residuals.

The test statistic is

\deqn{Q := n(n+2)\sum_{k=1}^{\ell} \frac{\hat{r}_k^2}{n-k}}

Where:

\deqn{\hat{r}_k := \frac{\sum_{i=k}^{n-1}(v_i-\bar{v})(v_{i-k} - \bar{v})}{\sum_{i=0}^{n-1}(v_i-\bar{v})^2}}

Where: \eqn{n} is the sample size (length of \eqn{v}) and \eqn{\ell} is the number of lags
}
\examples{
# Ljung-Box test for autocorrelation
ljung_box(c(1, 2, 3, 4, 5), lags = 2, fit_dof = 0)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/ljung_box.html}{Boost Documentation} for more details on the mathematical background.
}
