% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ljung_box_test.R
\name{ljung_box_test}
\alias{ljung_box_test}
\alias{ljung_box}
\title{Ljung-Box Test for Autocorrelation}
\usage{
ljung_box(v, lags = -1, fit_dof = 0)
}
\arguments{
\item{v}{A numeric vector.}

\item{lags}{A single integer value.}

\item{fit_dof}{A single integer value.}
}
\value{
A two-element numeric vector containing the test statistic and the p-value.
}
\description{
Functions to perform the Ljung-Box test for autocorrelation.
}
\examples{
# Ljung-Box test for autocorrelation
ljung_box(c(1, 2, 3, 4, 5), lags = 2, fit_dof = 0)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/ljung_box.html}{Boost Documentation} for more details on the mathematical background.
}
