% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_functions.R
\name{logistic_functions}
\alias{logistic_functions}
\alias{logistic_sigmoid}
\alias{logit}
\title{Logistic Functions}
\usage{
logistic_sigmoid(x)

logit(x)
}
\arguments{
\item{x}{Numeric value for which to compute the functions}
}
\value{
A single numeric value with the computed logit or logistic sigmoid function.
}
\description{
Functions to compute the logit and logistic sigmoid functions
}
\examples{
# Logistic Sigmoid Function
logistic_sigmoid(0.5)
# Logit Function
logit(0.7)
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/logistic.html}{Boost Documentation}
for more details on the mathematical background.
}
