% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_distribution.R
\name{normal_distribution}
\alias{normal_distribution}
\alias{normal_pdf}
\alias{normal_lpdf}
\alias{normal_cdf}
\alias{normal_lcdf}
\alias{normal_quantile}
\title{Normal Distribution Functions}
\usage{
normal_pdf(x, mean = 0, sd = 1)

normal_lpdf(x, mean = 0, sd = 1)

normal_cdf(x, mean = 0, sd = 1)

normal_lcdf(x, mean = 0, sd = 1)

normal_quantile(p, mean = 0, sd = 1)
}
\arguments{
\item{x}{quantile}

\item{mean}{mean parameter (default is 0)}

\item{sd}{standard deviation parameter (default is 1)}

\item{p}{probability (0 <= p <= 1)}
}
\description{
Functions to compute the probability density function, cumulative distribution function, and quantile function for the Normal distribution.
}
\examples{
# Normal distribution with mean = 0, sd = 1
normal_pdf(0)
normal_lpdf(0)
normal_cdf(0)
normal_lcdf(0)
normal_quantile(0.5)
}
\seealso{
\href{https://www.boost.org/doc/libs/1_87_0/libs/math/doc/html/math_toolkit/distributions/normal.html}{Boost Documentation} for more details on the mathematical background.
}
