% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/owens_t_function.R
\name{owens_t}
\alias{owens_t}
\title{Owens T Function}
\usage{
owens_t(h, a)
}
\arguments{
\item{h}{The first argument of the Owens T function}

\item{a}{The second argument of the Owens T function}
}
\value{
The value of the Owens T function at (h, a).
}
\description{
Computes the Owens T function of h and a, giving the probability of
the event (X > h and 0 < Y < a * X) where X and Y are independent
standard normal random variables.
}
