% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runs_tests.R
\name{runs_tests}
\alias{runs_tests}
\alias{runs_above_and_below_threshold}
\alias{runs_above_and_below_median}
\title{Runs Tests}
\usage{
runs_above_and_below_threshold(v, threshold)

runs_above_and_below_median(v)
}
\arguments{
\item{v}{A numeric vector.}

\item{threshold}{A single numeric value.}
}
\value{
A two-element numeric vector containing the t-statistic and the p-value.
}
\description{
Functions for Runs Tests.
}
\examples{
# Runs Above and Below Threshold
runs_above_and_below_threshold(c(1, 2, 3, 4, 5), threshold = 3)
# Runs Above and Below Median
runs_above_and_below_median(c(1, 2, 3, 4, 5))
}
\seealso{
\href{https://www.boost.org/doc/libs/latest/libs/math/doc/html/math_toolkit/runs_test.html}{Boost Documentation} for more details on the mathematical background.
}
