% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saspoint5_distribution.R
\name{saspoint5_distribution}
\alias{saspoint5_distribution}
\alias{saspoint5_pdf}
\alias{saspoint5_lpdf}
\alias{saspoint5_cdf}
\alias{saspoint5_lcdf}
\alias{saspoint5_quantile}
\title{S\eqn{\alpha}S Point5 Distribution Functions}
\usage{
saspoint5_pdf(x, location = 0, scale = 1)

saspoint5_lpdf(x, location = 0, scale = 1)

saspoint5_cdf(x, location = 0, scale = 1)

saspoint5_lcdf(x, location = 0, scale = 1)

saspoint5_quantile(p, location = 0, scale = 1)
}
\arguments{
\item{x}{quantile}

\item{location}{location parameter (default is 0)}

\item{scale}{scale parameter (default is 1)}

\item{p}{probability (0 <= p <= 1)}
}
\description{
Functions to compute the probability density function, cumulative distribution function, and quantile function for the S\eqn{\alpha}S Point5 distribution.
}
\examples{
Distribution only available with Boost version 1.87.0 or later.
if (packageVersion("BH") >= "1.87.0") {
# SaS Point5 distribution with location 0 and scale 1
  saspoint5_pdf(3)
  saspoint5_lpdf(3)
  saspoint5_cdf(3)
  saspoint5_lcdf(3)
  saspoint5_quantile(0.5)
}
}
\seealso{
\href{https://www.boost.org/doc/libs/1_87_0/libs/math/doc/html/math_toolkit/dist_ref/dists/saspoint5_dist.html}{Boost Documentation} for more details on the mathematical background.
}
