% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinus_cardinal_hyperbolic_functions.R
\name{sinus_cardinal_hyperbolic_functions}
\alias{sinus_cardinal_hyperbolic_functions}
\alias{sinc_pi}
\alias{sinhc_pi}
\title{Sinus Cardinal and Hyperbolic Functions}
\usage{
sinc_pi(x)

sinhc_pi(x)
}
\arguments{
\item{x}{Input value}
}
\value{
A single numeric value with the computed sinus cardinal or hyperbolic sinus cardinal function.
}
\description{
Functions to compute the sinus cardinal function and hyperbolic sinus cardinal function.
}
\examples{
# Sinus cardinal function
sinc_pi(0.5)
# Hyperbolic sinus cardinal function
sinhc_pi(0.5)
}
\seealso{
\href{https://www.boost.org/doc/libs/1_87_0/libs/math/doc/html/math_toolkit/sinc.html}{Boost Documentation} for more details on the mathematical background.
}
