% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_series.R
\name{tangent_t2n}
\alias{tangent_t2n}
\title{Tangent Numbers (\eqn{T_{2n}})}
\usage{
tangent_t2n(n)
}
\arguments{
\item{n}{Index of the tangent number (must be a non-negative integer)}
}
\value{
The value of the tangent number \eqn{\(T_{2n}\)}.
}
\description{
Tangent Numbers (\eqn{T_{2n}})
}
