% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemann_zeta_function.R
\name{zeta}
\alias{zeta}
\title{Riemann Zeta Function}
\usage{
zeta(z)
}
\arguments{
\item{z}{Real number input}
}
\value{
The value of the Riemann zeta function \eqn{(\zeta(z))}.
}
\description{
Computes the Riemann zeta function \eqn{(\zeta(s))} for argument \eqn{(z)}.
}
